% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAQILegend.R
\name{addAQILegend}
\alias{addAQILegend}
\title{Add an AQI Legend to a Map}
\usage{
addAQILegend(x = "topright", y = NULL, col = rev(AQI$colors),
  legend = rev(AQI$names), pch = 16, title = "Air Quality Index",
  ...)
}
\arguments{
\item{x}{x coordinate passed on to the \code{legend()} command}

\item{y}{y coordinate passed on to the \code{legend()} command}

\item{col}{the color for points/lines in the legend}

\item{legend}{a character vector to be shown in the legend}

\item{pch}{plotting symbols in the legend}

\item{title}{title for the legend}

\item{...}{additional arguments to be passed to \code{legend()}}
}
\description{
This function is a convenience wrapper around \code{graphics::legend()}. It will show the AQI colors and
names by default if \code{col} and \code{legend} are not specified.
}
\keyword{plotting}
