% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAQILines.R
\name{addAQILines}
\alias{addAQILines}
\title{Adds AQI Lines to a Plot}
\usage{
addAQILines(...)
}
\arguments{
\item{...}{additional arguments to be passed to \code{abline()}}
}
\description{
This function is a convenience for:

\code{abline(h=AQI$breaks_24, col=AQI$colors)}
}
\keyword{plotting}
