% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBullseye.R
\name{addBullseye}
\alias{addBullseye}
\title{Add a Bullseye at the Specified Location}
\usage{
addBullseye(lon, lat, map = NULL, cex = 2, lwd = 2)
}
\arguments{
\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{map}{optional RgoogleMaps map object}

\item{cex}{character expansion}

\item{lwd}{line width of individual circles}
}
\description{
Draws a bullseye with concentric circles of black and white.
}
\examples{
\dontrun{
# Google map
map <- monitorGoogleMap(ca)
addBullseye(ca$meta$longitude, ca$meta$latitude, map=map)
# line map
monitorMap(ca)
addBullseye(ca$meta$longitude, ca$meta$latitude)
}
}
\keyword{plotting}

