% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleMetadata.R
\name{addGoogleMetadata}
\alias{addGoogleMetadata}
\title{Add Elevation and Address Information to a Dataframe}
\usage{
addGoogleMetadata(df, lonVar = "longitude", latVar = "latitude")
}
\arguments{
\item{df}{dataframe with geolocation information (e.g. created by wrcc_qualityControl() or airsis_qualityControl)}

\item{lonVar}{name of longitude variable in the incoming dataframe}

\item{latVar}{name of the latitude variable in the incoming dataframe}
}
\value{
Input dataframe with additional columns: \code{elevation, siteName, countyName}.
}
\description{
Google APIs are used to determine elevation and
address information associated with the locations specified by the
\code{longitude} and \code{latitude} columns of the incoming dataframe.

Address information is obtained by using the \pkg{ggmap} package.
}
\references{
\url{https://maps.googleapis.com/maps/api/elevation}
}
\keyword{internal}

