% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIcon.R
\name{addIcon}
\alias{addIcon}
\title{Add an Icon to a RgoogleMap Plot
map, icon, lon, lat, expansion=0.1, pos=0}
\usage{
addIcon(map, icon, lon, lat, expansion = 0.1, pos = 0)
}
\arguments{
\item{map}{RgoogleMaps map object}

\item{icon}{object to be plotted}

\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{expansion}{icon expansion factor}

\item{pos}{position of icon relative to lon/lat (0=center, 1=bottom, 2=left, 3=top,4=right)}
}
\description{
Internal function aclled by monitorGoogleMap~ functions.
}

