% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addMazamaMetadata.R
\name{addMazamaMetadata}
\alias{addMazamaMetadata}
\title{Add Elevation and Address Information to a Dataframe}
\usage{
addMazamaMetadata(df, lonVar = "longitude", latVar = "latitude")
}
\arguments{
\item{df}{dataframe with geolocation information (e.g. created by wrcc_qualityControl() or airsis_qualityControl)}

\item{lonVar}{name of longitude variable in the incoming dataframe}

\item{latVar}{name of the latitude variable in the incoming dataframe}
}
\value{
Input dataframe with additional columns: timezone, countryCode, stateCode.
}
\description{
The \pkg{MazamaSpatialUtils} package used to determine elevation and
address information associated with the locations specified by the
\code{longitude} and \code{latitude} columns of the incoming dataframe.

This function requires previous setup of the \pkg{MazamaSpatialUtils} package:

\preformatted{
library(MazamaSpatialUtils)
setSpatialDataDir('~/Data/Spatial')
loadSpatialData('NaturalEarthAdm1')
}

If these lines have not be previously run, the function will return the original,
unmodified dataframe.
}
\references{
\url{https://github.com/MazamaScience/MazamaSpatialUtils}
}
\keyword{internal}

