% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addShadedBackground.R
\name{addShadedBackground}
\alias{addShadedBackground}
\title{Add Shaded Background}
\usage{
addShadedBackground(param, timeAxis, breaks = quantile(param, na.rm = TRUE),
  col = "blue", maxOpacity = 0.2, lwd = 1)
}
\arguments{
\item{param}{vector of data to be represented}

\item{timeAxis}{vector of times of the same length as param}

\item{breaks}{break methodology}

\item{col}{color for vertical lines}

\item{maxOpacity}{maximum opacity}

\item{lwd}{line width}
}
\description{
Adds vertical lines to an existing plot using any variable that shares the same
}

