% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addShadedNight.R
\name{addShadedNight}
\alias{addShadedNight}
\title{Add Nighttime Shading to a Plot}
\usage{
addShadedNight(timeInfo, col = adjustcolor("black", 0.1))
}
\arguments{
\item{timeInfo}{dataframe with local time, sunrise, and sunset}

\item{col}{color used to shade nights -- defaults to \code{adjustcolor('black',0.2)}}
}
\description{
Draw shading rectangles on a plot to indicate nighttime hours.
}
\seealso{
\link{timeInfo}
}
