% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addShadedNights.R
\name{addShadedNights}
\alias{addShadedNights}
\title{Add Nighttime Shading on an Existing Plot}
\usage{
addShadedNights(timeInfo, col = adjustcolor("black", 0.1))
}
\arguments{
\item{timeInfo}{dataframe with local time, sunrise, and sunset}

\item{col}{color used to shade nights -- defaults to \code{adjustcolor('black',0.2)}}
}
\description{
Draw shading rectangles on a plot to indicate nighttime hours.
}
\seealso{
\link{timeInfo}
}

