% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUSGSElevation.R
\name{addUSGSElevation}
\alias{addUSGSElevation}
\title{Add Elevation Data to a Dataframe}
\usage{
addUSGSElevation(df, lonVar = "longitude", latVar = "latitude",
  existingMeta = NULL)
}
\arguments{
\item{df}{dataframe with geolocation information (\emph{e.g.} created by \code{wrcc_qualityControl()} or \code{airsis_qualityControl})}

\item{lonVar}{name of longitude variable in the incoming dataframe}

\item{latVar}{name of the latitude variable in the incoming dataframe}

\item{existingMeta}{existing 'meta' dataframe from which to obtain metadata for known monitor deployments}
}
\value{
Input dataframe with (possibly) additional column: \code{elevation}.
}
\description{
USGS APIs are used to determine elevation associated with the locations specified by the
\code{lonVar} and \code{latVar} columns of the incoming dataframe.
}
\references{
\url{https://nationalmap.gov/epqs/}
}
\keyword{internal}
