% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustRange.R
\name{adjustRange}
\alias{adjustRange}
\title{Create a Scaled Range}
\usage{
adjustRange(x, factor = 1.1, minRange = 0.1)
}
\arguments{
\item{x}{vector of numeric values}

\item{factor}{scaling factor to be applied}

\item{minRange}{minimum range to be returned}
}
\value{
Vector of low and high values of the adjusted range.
}
\description{
Scale the range of values by a given factor. The min and max values of the 
original range of valus are adjusted so that the new range is larger or smaller by \code{factor}.
The returned range will always be equal to or greater than \code{minRange}.
}
\examples{
\dontrun{
x <- (1:10)
range(x)
[1] 1 10
x_new <- adjustRange(x, factor = 2)
range(x_new)
[1] -3.5 14.5
}
}
\keyword{internal}

