% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_createLatestDataDataframes.R
\name{airnow_createLatestDataDataframes}
\alias{airnow_createLatestDataDataframes}
\title{Return Reshaped, Monthly Dataframes of the Latest AirNow Data}
\usage{
airnow_createLatestDataDataframes(user, pass, parameters = NULL, tries = 6,
  verbose = FALSE)
}
\arguments{
\item{user}{user name}

\item{pass}{pass}

\item{parameters}{vector of names of desired pollutants or NULL for all pollutants}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical requesting verbose output from libcurl}
}
\value{
Returns a list of dataframes where each dataframe contains all data for a unique parameter (e.g: "PM2.5", "NOX").
}
\description{
The airnow_createLatestDataDataframes() function uses the \link{airnow_downloadLatestData} function 
to download dataframes of the most recent AirNow data and restructures that data into a format that is compatible
with the PWFSLSmoke package \emph{ws_monitor} data model.

AirNow data parameters include at least the following list:
\enumerate{
\item{BARPR}
\item{BC}
\item{CO}
\item{NO}
\item{NO2}
\item{NO2Y}
\item{NO2X}
\item{NOX}
\item{NOOY}
\item{OC}
\item{OZONE}
\item{PM10}
\item{PM2.5}
\item{PRECIP}
\item{RHUM}
\item{SO2}
\item{SRAD}
\item{TEMP}
\item{UV-AETH}
\item{WD}
\item{WS}
}

Setting \code{parameters=NULL} will generate a separate dataframe for each of the above parameters.
}
\examples{
\dontrun{
airnow_latestData <- airnow_createLatestDataDataframe(user, pass, "PM2.5")
}
}
\seealso{
\link{airnow_downloadLatestData}

\link{airnow_qualityControl}
}
\keyword{AirNow}

