% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airnow_createMetaDataframe.R
\name{airnow_createMetaDataframe}
\alias{airnow_createMetaDataframe}
\title{Create Sites Metadata Dataframe}
\usage{
airnow_createMetaDataframe(sites, parameter = "PM2.5", verbose = FALSE)
}
\arguments{
\item{sites}{dataframe obtained from airnow_downloadSites()}

\item{parameter}{parameter name used to subset sites}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Dataframe of site location metadata appropriate for use in the
creation of a \code{ws_monitor} object.
}
\description{
The \pkg{PWFSLSmoke} package's data model for monitoring
data is a list with \code{data} and \code{meta} dataframes. The \code{meta}
dataframe has metadata associated with monitoring site locations for a specific
parameter and must contain at least the following columns:
\itemize{
  \item{monitorID -- per deployment unique ID}
  \item{longitude -- decimal degrees E}
  \item{latitude -- decimal degrees N}
  \item{elevation -- height above sea level in meters}
  \item{timezone -- olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}

The \code{meta} dataframe will have rownames matching \code{monitorID}.

The airnow_downloadSites() function takes a dataframe obtained from AirNowTech's
\code{monitoring_site_locations.dat} file and performs the following cleanup:
\itemize{
  \item{remove duplicate entries associated with different parameters}
  \item{convert incorrect values to \code{NA} e.g. lon=0 & lat=0}
  \item{add timezone information}
}

Metadata subsets can be generated for sites monitoring specific pollutants
by specifying, for example \code{parameter='O3'}.
}
\examples{
\dontrun{
sites <- airnow_downloadSites('USER','PASS')
meta <- airnow_createMetaDataframe(sites, parameter='PM2.5')
}
}
\seealso{
\link{airnow_downloadSites}
}
\keyword{AirNow}

