% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airnow_downloadData.R
\name{airnow_downloadData}
\alias{airnow_downloadData}
\title{Download and Aggregate Multiple Hourly Data Files from AirNow}
\usage{
airnow_downloadData(user = "", pass = "", startdate = "", hours = 24,
  tries = 6, verbose = FALSE, curl = NULL)
}
\arguments{
\item{user}{user name}

\item{pass}{password}

\item{startdate}{desired staring date (integer or character representing YYYYMMDDHH)}

\item{hours}{desired number of hours of data to assemble}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical flag to generate verbose output}

\item{curl}{previously initialized CURL context/handle (see RCurl::getURL())}
}
\value{
Dataframe of aggregated AirNow data.
}
\description{
This function makes repeated calls to \link{airnow_downloadHourlyData}
to obtain data from AirNow. A single curl handle is created and reused to improve
peformance and avoid errors like "Maximum login connections". All data obtained are then
combined into a single dataframe and returned.
}
\examples{
\dontrun{
df <- airnow_downloadData(USER, PASS, 2015070112, hours=48)
}
}
\seealso{
\link{airnow_downloadMonthlyData}
}
\keyword{AirNow}

