% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_downloadHourlyData.R
\name{airnow_downloadHourlyData}
\alias{airnow_downloadHourlyData}
\title{Download Hourly Data from AirNow}
\usage{
airnow_downloadHourlyData(datestamp = "",
  baseUrl = "https://files.airnowtech.org/airnow")
}
\arguments{
\item{datestamp}{integer or character representing YYYYMMDDHH}

\item{baseUrl}{base URL for archived hourly data}

\item{tries}{number of download attempts in the face of timeouts}

\item{curl}{previously initialized CURL context/handle (see RCurl::getURL())}
}
\value{
Dataframe of AirNow hourly data.
}
\description{
The \url{http://airnowtech.org} site provides both air pollution
monitoring data as well as monitoring site location metadata. The \code{airnow_downloadHourlyData()}
function retrieves a single, hourly data file and returns it as a dataframe.
}
\note{
:  As of 2016-12-27, it appears that hourly data are available only for 2016 and
not for earlier years.
}
\examples{
\dontrun{
df <- airnow_downloadHourlyData(2016070112)
}
}
\seealso{
\link{airnow_createDataDataframes}

\link{airnow_downloadData}
}
\keyword{AirNow}
\keyword{internal}

