% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_loadLatest.R
\name{airnow_loadLatest}
\alias{airnow_loadLatest}
\title{Load Recent AirNow Monitoring Data}
\usage{
airnow_loadLatest(baseUrl = "https://haze.airfire.org/monitoring/RData/",
  file = "airnow_pm25_latest.RData")
}
\arguments{
\item{baseUrl}{location of the AirNow latest data file}

\item{file}{name of the AirNow latest data file}
}
\value{
A \emph{ws_monitor} object with AirNow data.
}
\description{
The most recent 45 days of AirNow data are updated in real time
at PWFSL and can be loaded with this function.
}
\examples{
\dontrun{
airnow <- airnow_loadLatest()
}
}
\keyword{AirNow}
