% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{AIRSIS}
\alias{AIRSIS}
\title{AIRSIS Unit Types}
\format{A list of lists}
\usage{
AIRSIS
}
\description{
AIRSIS provides access to data by unit type at URLs like:
  http://usfs.airsis.com/vision/common/CSVExport.aspx?utid=38&StartDate=2017-11-06&EndDate=2017-11-07

Unit types include:
\itemize{
\item{DATARAM}{ 21 = Dataram}
\item{BAM1020}{ 24 = Bam 1020}
\item{EBAM_NEW}{ 30 = eBam-New}
\item{EBAM}{ 38 = Iridium - Ebam}
\item{ESAM}{ 39 = Iridium - Esam}
\item{AUTOMET}{ 43 = Automet}
}
}
\details{
AIRSIS monitor types and codes
}
\note{
This list of monitors was created on Feb 09, 2017.
}
\keyword{datasets}
