% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_buildLibrary.R
\name{airsis_buildLibrary}
\alias{airsis_buildLibrary}
\title{Build a Local RData Library from AIRSIS}
\usage{
airsis_buildLibrary(provider = "USFS", unitIDs = NULL, years = 2000:2015,
  logLevel = futile.logger::INFO, transcript = "AIRSIS_TRANSCRIPT.txt",
  outputDir = getwd(),
  baseUrl = "http://xxxx.airsis.com/vision/common/CSVExport.aspx?")
}
\arguments{
\item{provider}{identifier used to modify baseURL \code{['APCD'|'USFS']}}

\item{unitIDs}{vector of character or numeric unit identifiers}

\item{years}{vector of years for which to download data}

\item{logLevel}{log level for output \code{['TRACE','DEBUG','INFO','WARNING','ERROR','FATAL']}}

\item{transcript}{filename for logging output}

\item{outputDir}{directory where data files and transcript will be written}

\item{baseUrl}{base URL for data queries}
}
\value{
Vector of file names created.
}
\description{
Smoke monitoring data from airsis.com is downloaded, quality controlled
and converted in to a collection of .RData files, each containing a \code{ws_monitor}
for a particular UnitID.

By default, a transcript file will be generated containing information on the progress
of data downlaod, quality control and restructuring. More detailed information is
generated by setting \code{logLevel='DEBUG'}.

Setting \code{transcript=NULL} causes output to be sent to the console.
}
\details{
When no unitIDs are supplied (the default), a range of unitIDs is generated that
covers all AIRSIS monitor known to exist for each provider as of May, 2016. The range will
contain some unitIDs for which no monitors exist so it is normal to see log messages
for unitIDs with no data.
}
\examples{
\dontrun{
files <- airsis_buildLibrary('APCD', years=2010:2015, outputDir='~/Data/AIRSIS')
monitor_combineLibraryFiles('~/Data/AIRSIS', files, 'AIRSIS_APCD.RData')
apcd_CA <- <- airsis_load(stateCodes='CA', url='~/Data/AIRSIS/AIRSIS_APCD.RData')
monitor_leaflet(apcd_CA)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\seealso{
\link{monitor_combineLibraryFiles}
}
\keyword{AIRSIS}

