% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_createDataDataframe.R
\name{airsis_createDataDataframe}
\alias{airsis_createDataDataframe}
\title{Create Data Dataframe}
\usage{
airsis_createDataDataframe(df, meta, pm25Var = "ConcHr", verbose = FALSE)
}
\arguments{
\item{df}{single site AIRSIS dataframe created by airsis_clustering()}

\item{meta}{AIRSIS meta dataframe created by airsis_createMetaDataframe()}

\item{pm25Var}{name of PM 2.5 variable in the incoming dataframe}

\item{verbose}{logical flag to generate verbose output}
}
\value{
ws_monitor object with a unique `monitorID` for each unique location
}
\description{
Explain this function
}
\keyword{AIRSIS}

