% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_createMetaDataframe.R
\name{airsis_createMetaDataframe}
\alias{airsis_createMetaDataframe}
\title{Create Sites Metadata Dataframe}
\usage{
airsis_createMetaDataframe(df, verbose = FALSE)
}
\arguments{
\item{df}{single site AIRSIS dataframe after clustering}

\item{verbose}{logical flag to generate verbose output}
}
\value{
ws_monitor object with a unique `monitorID` for each unique location
}
\description{
Explain this function
}
\seealso{
\link{airsis_downloadData}

\link{addClustering}
}
\keyword{AIRSIS}

