% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_createMonitorObject.R
\name{airsis_createMonitorObject}
\alias{airsis_createMonitorObject}
\title{Convert AIRSIS Data into a ws_monitor Object}
\usage{
airsis_createMonitorObject(df, clusterDiameter = 1000, verbose = FALSE)
}
\arguments{
\item{df}{single site dataframe created by airsis_downloadData()}

\item{clusterDiameter}{diameter in meters used to determine the number of clusters (see description)}

\item{verbose}{logical flag to generate verbose output}
}
\value{
ws_monitor object with a unique `monitorID` for each unique location
}
\description{
Convert a AIRSIS dataframe into a ws_monitor object, ready for use
with all monitor_~ functions. Steps involved include:

\enumerate{
 \item{quality control}
 \item{clustering to determine unique deployments}
 \item{metadata enhancement including: elevation, timezone, state, country, site name}
 \item{reshaping data into a time-by-deployment dataframe}
}
}
\keyword{AIRSIS}

