% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_createRawDataframe.R
\name{airsis_createRawDataframe}
\alias{airsis_createRawDataframe}
\title{Obtain AIRSIS Data and Parse into a Dataframe}
\usage{
airsis_createRawDataframe(provider = "USFS", unitID = NULL,
  startdate = 20020101, enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz
  = "GMT"), clusterDiameter = 1000,
  baseUrl = "http://xxxx.airsis.com/vision/common/CSVExport.aspx?",
  saveFile = NULL)
}
\arguments{
\item{provider}{identifier used to modify baseURL \code{['APCD'|'USFS']}}

\item{unitID}{character or numeric AIRSIS unit identifier}

\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{clusterDiameter}{diameter in meters used to determine the number of clusters (see \code{addClustering})}

\item{baseUrl}{base URL for data queries}

\item{saveFile}{optional filename where raw CSV will be written}
}
\value{
ws_monitor object with a unique `monitorID` for each unique deployment.
}
\description{
Obtains monitor data from an AIRSIS webservice and converts
it into a quality controlled, metadata enhanced "raw" dataframe
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download CSV text}
 \item{parse CSV text}
 \item{apply quality control}
 \item{apply clustering to determine unique deployments}
 \item{enhance metadata to include: elevation, timezone, state, country, site name}
}
}
\note{
The downloaded CSV may be saved to a local file by providing an argument to the \code{saveFile} parameter.
}
\seealso{
\code{\link{airsis_downloadData}}

\code{\link{airsis_parseData}}

\code{\link{airsis_qualityControl}}

\code{\link{addClustering}}
}
\keyword{AIRSIS}

