% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_currentData.R
\name{airsis_currentData}
\alias{airsis_currentData}
\title{Load Current Airsis Data}
\usage{
airsis_currentData(lookbackDays = 7, providers = c("usfs", "apcd", "arb2",
  "epa"), usernames, passwords)
}
\arguments{
\item{lookbackDays}{number of days prior to now considered to be "current"}

\item{providers}{vector of data providers}

\item{usernames}{vector of usernames corresponding to \code{providers}}

\item{passwords}{vector of passwords corresponding to \code{providers}}
}
\value{
\code{ws_monitor} object with data for monitors with new data in last \code{lookbackDays} days
}
