% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_currentStatus.R
\name{airsis_currentStatus}
\alias{airsis_currentStatus}
\title{Load Current Status from Airsis Site}
\usage{
airsis_currentStatus(provider, username, password)
}
\arguments{
\item{provider}{data provider, e.g. usfs, arb2, etc.}

\item{username}{username for login credentials}

\item{password}{password for login credentials}
}
\value{
A dataframe with unit ID, alias, location, and latest update time, in UTC.
}
\description{
Scrape current monitor status from Airsis website (http://xxxx.airsis.com
where xxxx %in% c("usfs", "apcd", "arb2", "epa").
}
\examples{
\dontrun{
usfs <- airsis_currentStatus(provider="usfs")
View(usfs)
}
}
