% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_downloadData.R
\name{airsis_downloadData}
\alias{airsis_downloadData}
\title{Download USFS Data from AIRSIS}
\usage{
airsis_downloadData(unitID = NULL, startdate = 20020101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  baseUrl = "http://usfs.airsis.com/vision/common/CSVExport.aspx?")
}
\arguments{
\item{unitID}{unit identifier}

\item{startdate}{integer or character representing start date as YYYYMMDD (GMT)}

\item{enddate}{integer or character representing end date as YYYYMMDD (GMT)}

\item{baseUrl}{base URL for data queries}
}
\value{
Dataframe of AIRSIS monitor data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a dataframe.
}
\examples{
\dontrun{
df <- airsis_downloadData('1026',startdate=20150701,enddate=201231)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

