% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_EBAMQualityControl.R
\name{airsis_EBAMQualityControl}
\alias{airsis_EBAMQualityControl}
\title{Apply Quality Control to raw AIRSIS EBAM tibble}
\usage{
airsis_EBAMQualityControl(tbl, valid_Longitude = c(-180, 180),
  valid_Latitude = c(-90, 90), remove_Lon_zero = TRUE,
  remove_Lat_zero = TRUE, valid_Flow = c(16.7 * 0.95, 16.7 * 1.05),
  valid_AT = c(-Inf, 45), valid_RHi = c(-Inf, 45),
  valid_Conc = c(-Inf, 5), flagAndKeep = FALSE)
}
\arguments{
\item{tbl}{single site tibble created by \code{airsis_parseData()}}

\item{valid_Longitude}{range of valid Longitude values}

\item{valid_Latitude}{range of valid Latitude values}

\item{remove_Lon_zero}{flag to remove rows where Longitude == 0}

\item{remove_Lat_zero}{flag to remove rows where Latitude == 0}

\item{valid_Flow}{range of valid Flow values}

\item{valid_AT}{range of valid AT values}

\item{valid_RHi}{range of valid RHi values}

\item{valid_Conc}{range of valid ConcHr values}

\item{flagAndKeep}{flag, rather than remove, bad data during the QC process}
}
\value{
Cleaned up tibble of AIRSIS monitor data.
}
\description{
Perform various QC measures on AIRSIS EBAM data.

The following columns of data are tested against valid ranges:
\itemize{
\item{\code{Flow}}
\item{\code{AT}}
\item{\code{RHi}}
\item{\code{ConcHr}}
}

A \code{POSIXct datetime} column (UTC) is also added based on \code{Date.Time.GMT}.
}
\seealso{
\code{\link{airsis_qualityControl}}
}
\keyword{AIRSIS}
