% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_fileString.R
\name{airsis_fileString}
\alias{airsis_fileString}
\title{Download Data from AIRSIS}
\usage{
airsis_fileString(provider = "USFS", unitID = NULL, startdate = 20020101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  baseUrl = "http://xxxx.airsis.com/vision/common/CSVExport.aspx?")
}
\arguments{
\item{provider}{identifier used to modify baseURL \code{['APCD'|'USFS']}}

\item{unitID}{unit identifier}

\item{startdate}{integer or character representing start date as YYYYMMDD (GMT)}

\item{enddate}{integer or character representing end date as YYYYMMDD (GMT)}

\item{baseUrl}{base URL for data queries}
}
\value{
String containing AIRSIS output.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the AIRIS output.
}
\examples{
\dontrun{
fileString <- airsis_downloadData(USFS',unitID='1026',startdate=20150701,enddate=20151231)
df <- airsis_parseData(fileString)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

