% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_loadLatest.R
\name{airsis_loadLatest}
\alias{airsis_loadLatest}
\title{Load Recent AIRSIS Monitoring Data}
\usage{
airsis_loadLatest(baseUrl = "https://haze.airfire.org/monitoring/RData/",
  file = "airsis_pm25_latest.RData")
}
\arguments{
\item{baseUrl}{location of the AIRSIS latest data file}

\item{file}{name of the AIRSIS latest data file}
}
\value{
A \emph{ws_monitor} object with AIRSIS data.
}
\description{
The most recent 45 days of AIRSIS data are updated in real time
at PWFSL and can be loaded with this function.
}
\examples{
\dontrun{
airsis <- airsis_loadLatest()
}
}
\keyword{AIRSIS}
