% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_parseData.R
\name{airsis_parseData}
\alias{airsis_parseData}
\title{Parse AIRSIS Data String}
\usage{
airsis_parseData(fileString)
}
\arguments{
\item{fileString}{character string containing AIRSIS data as a csv}
}
\value{
Dataframe of AIRSIS monitor data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a dataframe.
}
\examples{
\dontrun{
fileString <- airsis_downloadData(USFS',unitID='1026',startdate=20150701,enddate=20151231)
df <- airsis_parseData(fileString)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

