% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_qualityControl.R
\name{airsis_qualityControl}
\alias{airsis_qualityControl}
\title{Apply Quality Control to Raw AIRSIS Dataframe}
\usage{
airsis_qualityControl(df, verbose = FALSE)
}
\arguments{
\item{df}{single site dataframe created by airsis_downloadData()}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Cleaned up dataframe of AIRSIS monitor data.
}
\description{
Perform various QC measures on AIRSIS data.
}
\keyword{AIRSIS}

