% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_readData.R
\name{airsis_readData}
\alias{airsis_readData}
\title{Read and Parse Single Monitor Data from AIRSIS}
\usage{
airsis_readData(file, monitorName = NULL, verbose = FALSE)
}
\arguments{
\item{file}{either a path to a AIRSIS data file the contents of the file as a character string}

\item{monitorName}{human readable monitor identifier}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Dataframe of AIRSIS monitor data.
}
\description{
An AIRSIS monitor data file is read parsed and the results are returned as a dataframe.
}
\references{
\href{http://app.airsis.com/usfs/}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

