% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsisDump_parseData.R
\name{airsisDump_parseData}
\alias{airsisDump_parseData}
\title{Parse AIRSIS Dump File Data String}
\usage{
airsisDump_parseData(fileString)
}
\arguments{
\item{fileString}{character string containing AIRSIS dump file}
}
\value{
List of dataframes of WRCC raw monitor data from multiple monitors.
}
\description{
Raw character data from AIRSIS are parsed into a dataframe.
The incoming \code{fileString} can be read from a local
file using \code{readr::read_file()}.

This function is intended for internal use at the Pacific Wildland
Fire Sciences Lab.
}
\keyword{internal}

