% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{AQI_en}
\alias{AQI_en}
\title{Official Air Quality Index Levels, Names and Colors}
\format{
A list with named elements
}
\usage{
AQI_en
}
\description{
Official AQI levels, names and colors are provided in a list for easy coloring and labeling.
}
\details{
AQI breaks and associated names and colors (english language)


The \code{AQI_es} object contains english language text. It is equalivalent to the
\code{AQI} object and provided for consistency with other language versions.

AQI breaks and colors are defined at
\url{https://docs.airnowapi.org/aq101}
}
\note{
The low end of each break category is used as the breakpoint.
}
\seealso{
\code{\link{AQI}} \code{\link{AQI_es}}
}
\keyword{datasets}
