% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{Camp_Fire}
\alias{Camp_Fire}
\title{Camp Fire example dataset}
\format{
A \emph{ws_monitor} object with "meta" and "data" dataframes.
}
\usage{
Camp_Fire
}
\description{
The \code{Camp_Fire} dataset provides a quickly loadable
version of a \emph{ws_monitor} object for practicing and code examples.

This dataset was was generated on 2021-11-18 by running:

\preformatted{
library(PWFSLSmoke)

Camp_Fire <-
  monitor_loadAnnual(2018) %>%
  monitor_subset(stateCodes = 'CA') %>%
  monitor_subset(tlim = c(20181108, 20181123))

save(Camp_Fire, file = "data/Camp_Fire.RData")
}
}
\keyword{datasets}
