% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{Carmel_Valley}
\alias{Carmel_Valley}
\title{Carmel Valley example dataset}
\format{
A \emph{ws_monitor} object with "meta" and "data" dataframes.
}
\usage{
Carmel_Valley
}
\description{
In August of 2016, the Soberanes fire in California burned along the Big Sur
coast. It was at the time the most expensive wildfire in US history. This dataset contains
PM2.5 monitoring data for the monitor in Carmel Valley which shows heavy smoke
as well as strong diurnal cycles associated with sea breezes.

The \code{Camp_Fire} dataset provides a quickly loadable
version of a \emph{ws_monitor} object for practicing and code examples.

documentation.
}
\keyword{datasets}
