% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyMonitor.R
\name{createEmptyMonitor}
\alias{createEmptyMonitor}
\title{Create empty monitor object}
\usage{
createEmptyMonitor()
}
\value{
A empty \emph{ws_monitor} object.
}
\description{
The list of columns in the returned \code{meta} dataframe is:

\preformatted{
> names(meta)
 [1] "monitorID"             "longitude"             "latitude"
 [4] "elevation"             "timezone"              "countryCode"
 [7] "stateCode"             "siteName"              "agencyName"
[10] "countyName"            "msaName"               "monitorType"
[13] "monitorInstrument"     "aqsID"                 "pwfslID"
[16] "pwfslDataIngestSource" "telemetryAggregator"   "telemetryUnitID"
}
}
\examples{
emptyMonitor <- createEmptyMonitor()
monitor_isMonitor(emptyMonitor)
monitor_isEmpty(emptyMonitor)
}
\keyword{internal}
