% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_createMetaDataframe.R
\name{epa_createMetaDataframe}
\alias{epa_createMetaDataframe}
\title{Create Sites Metadata Dataframe}
\usage{
epa_createMetaDataframe(df, verbose)
}
\arguments{
\item{df}{an EPA dataframe after metadata enhancement}
}
\value{
A \code{'meta'} dataframe for use in a \code{ws_monitor} object.
}
\keyword{ws_monitor}

