% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{esriToken}
\alias{esriToken}
\title{Token used for ESRI Geocoding Requests}
\format{Character string.}
\description{
All package functions that interact with ESRI location services will
use the token whenever a request is made.
}
\seealso{
addEsriAddress
}
\keyword{environment}
