% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_downloadData.R
\name{generic_downloadData}
\alias{generic_downloadData}
\title{Download generic data}
\usage{
generic_downloadData(filePath)
}
\arguments{
\item{filePath}{Either a path to a file, or a connection (\code{http(s)://},
\code{ftp(s)://}).}
}
\value{
A character vector of length 1, containing data from the file located
  at \code{filePath}.
}
\description{
This function takes a location to a delimited file, gets the file, and
returns a string containing the file data.
}
\details{
This function is essentailly a wrapper for \code{\link[readr]{read_file}}.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

# make current directory PWFSLSmoke package directory
filePath <- "./localData/airsis_ebam_example-clean.csv"

fileString <- generic_downloadData(filePath)

}, silent = FALSE)
}
}
