% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{getEsriToken}
\alias{getEsriToken}
\title{Get ESRI Token}
\usage{
getEsriToken()
}
\value{
String.
}
\description{
Returns the current esriToken.
}
\seealso{
addEsriAddress

esriToken

setEsriToken
}
\keyword{environment}
