% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{getGoogleApiKey}
\alias{getGoogleApiKey}
\title{Get Google API Key}
\usage{
getGoogleApiKey()
}
\value{
String.
}
\description{
Returns the current Google API key.
}
\seealso{
addGoogleAddress

addGoogleElevation

googleApiKey

setGoogleApiKey
}
\keyword{environment}
