% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{googleApiKey}
\alias{googleApiKey}
\title{API Key used for Google Geocoding Requests}
\format{
Character string.
}
\description{
All package functions that interact with Google location services will
use API key whenever a request is made.
}
\seealso{
addGoogleAddress

addGoogleElevation
}
\keyword{environment}
