% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load PM2.5 monitoring data}
\usage{
loadData(startdate = NULL, enddate = NULL, monitorIDs = NULL,
  latestDataDir = NULL, dailyDataDir = NULL, annualDataDir = NULL)
}
\arguments{
\item{startdate}{Desired start date (integer or character in ymd[hms] format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in ymd[hms] format
or \code{POSIXct}).}

\item{monitorIDs}{Optional vector of monitorIDs.}

\item{latestDataDir}{Local Directory where 'latest' data files are found.}

\item{dailyDataDir}{Local Directory where 'daily' data files are found.}

\item{annualDataDir}{Local Directory where annual data files are found.}
}
\value{
A \emph{ws_monitor} object with PM2.5 monitoring data.
}
\description{
Loads monitoring data for a given time range. Data from AirNow,
AIRSIS and WRCC are combined into a single \emph{ws_monitor} object.

Archival datasets are joined with 'daily' and 'latest' datasets as needed to
satisfy the requested date range.
}
\note{
Joining datasets is a computationally expensive task when many monitors
are involved. It is highly recommend that \code{monitorIDs} be specified when
loading recent data with this function.
}
\examples{
\dontrun{
ca <- loadData(20170601,20171001) \%>\% monitor_subset(stateCodes='CA')
}
}
\seealso{
\code{\link{loadDaily}}

\code{\link{loadLatest}}
}
\keyword{AIRSIS}
\keyword{AirNow}
\keyword{WRCC}
