% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFile.R
\name{loadDataFile}
\alias{loadDataFile}
\title{Load data from URL or local file}
\usage{
loadDataFile(
  filename = NULL,
  dataUrl = "https://haze.airfire.org/monitoring/latest/RData",
  dataDir = NULL
)
}
\arguments{
\item{filename}{Name of the data file to be loaded.}

\item{dataUrl}{Remote URL directory for data files.}

\item{dataDir}{Local directory containing data files.}

\item{baseUrl}{Base URL for data files.}
}
\value{
A \emph{ws_monitor} object.

A data object.
}
\description{
Loads pre-generated .RData files

Loads pre-generated .rda files from a URL or a local
directory. This function is intended to be called by other \code{~_load()}
functions and can remove internet latencies when local versions of daata are
available.

For this reason, specification of \code{dataDir} always takes precedence over
\code{dataUrl}.
}
\keyword{internal}
