% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_aqi.R
\name{monitor_aqi}
\alias{monitor_aqi}
\title{Calculate hourly AQI values}
\usage{
monitor_aqi(ws_monitor, aqiParameter = "pm25", nowcastVersion = "pm",
  includeShortTerm = FALSE)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}

\item{aqiParameter}{parameter type}

\item{nowcastVersion}{character identity specifying the type of nowcast algorithm to be used}

\item{includeShortTerm}{calcluate preliminary values starting with the 2nd hour}
}
\description{
Nowcast and AQI algorithms are applied to the data in in the ws_monitor object.
}
\references{
\url{https://www3.epa.gov/airnow/aqi-technical-assistance-document-may2016.pdf}

\url{https://archive.epa.gov/ttn/ozone/web/pdf/airqual.pdf}

\url{https://www.ecfr.gov/cgi-bin/retrieveECFR?n=40y6.0.1.1.6#se40.6.58_150} (see Appendix G)
}
\keyword{ws_monitor}
