% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_collapse.R
\name{monitor_collapse}
\alias{monitor_collapse}
\title{Collapse a ws_monitor Object into a ws_monitor Object with a Single Monitor}
\usage{
monitor_collapse(ws_monitor, lon = NULL, lat = NULL,
  monitorID = "generated_id", FUN = mean, na.rm = TRUE, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{lon}{longitude of the collapsed monitoring station. (Default = mean of the longitudes)}

\item{lat}{latitude of the collapsed monitoring station. (Default = mean of the latitudes)}

\item{monitorID}{monitor ID of the collapsed monitoring station.}

\item{FUN}{function to be applied to all the monitors at a single time index.}

\item{na.rm}{logical value indicating whether NA values should be ignored when FUN is applied}

\item{...}{additional arguments to be passed on to the apply() function}
}
\value{
A ws_monitor object with meta and data that corresponds to the collapsed single monitor
}
\description{
Collapses data from all the monitors in \code{ws_monitor} into a single-monitor 
ws_monitor object using the function provided in the \code{FUN} argument. The single-monitor
result will be located at the mean longitude and latitude unless \code{lon} and \code{lat}
parameters are specified.
}
\keyword{ws_monitor}

