% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_combine.R
\name{monitor_combine}
\alias{monitor_combine}
\title{Combine ws_monitor Objects}
\usage{
monitor_combine(monitorList)
}
\arguments{
\item{monitorList}{list containing \code{ws_monitor} objects}
}
\value{
A ws_monitor object.
}
\description{
Given a list of \code{ws_monitor} objects, the function generates a single ws_monitor object
by merging the 'meta' and 'data' dataframes from each object.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke/')
monitorList <- list()
monitorList[[1]] <- airsis_createMonitorObject('USFS', '1031', 20160701, 20161231)
monitorList[[2]] <- airsis_createMonitorObject('USFS', '1032', 20160701, 20161231)
monitorList[[3]] <- airsis_createMonitorObject('USFS', '1033', 20160701, 20161231)
monitorList[[4]] <- airsis_createMonitorObject('USFS', '1034', 20160701, 20161231)
ws_monitor <- monitor_combine(monitorList)
monitor_dygraph(ws_monitor)
}
}
\keyword{ws_monitor}

