% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_combineLibraryFiles.R
\name{monitor_combineLibraryFiles}
\alias{monitor_combineLibraryFiles}
\title{Combine Individual Monitor Files}
\usage{
monitor_combineLibraryFiles(outputDir = getwd(), inputFiles = NULL,
  outputFile = NULL)
}
\arguments{
\item{outputDir}{directory where data files are located}

\item{inputFiles}{vector of file names to be ingested}

\item{outputFile}{name of combined file}
}
\value{
Name of combined file.
}
\description{
Load and combine .RData files containing \code{ws_monitor} objects.

When \code{inputFiles=NULL} (the default), all .RData files are used.
}
\examples{
\dontrun{
files <- airsis_buildLibrary('APCD', years=2010:2015, outputDir='~/Data/AIRSIS')
monitor_combineLibraryFiles('~/Data/AIRSIS', files, 'AIRSIS_APCD.RData')
apcd_CA <- <- airsis_load(stateCodes='CA', url='~/Data/AIRSIS/AIRSIS_APCD.RData')
monitor_leaflet(apcd_CA)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\seealso{
\link{airsis_buildLibrary}
}
\keyword{AIRSIS}

