% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_currentData.R
\name{monitor_currentData}
\alias{monitor_currentData}
\title{Create a dataframe of current monitor data}
\usage{
monitor_currentData(ws_monitor,
  monitoringUrlBase = "http://tools.airfire.org/monitoring/v4/#/?monitors=")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
A tibble of "lastValid" data and associated timing information.
}
\details{
Data are assigned to the beginning of the hour they represent. So a PM2.5 value assigned to 2pm
will represent data averaged over the period 14:00 - 14:59. This is in keeping with a day representing
00:00 - 23:59 and a month representing the beginning of the month until the end of the month.

Because of this, data for 2pm should never be available until just after 3pm. If it is currently 3:15pm
then we need to subtract 1 hour from the current \code{processingTime} before subtracting the \code{lastValidTime}
to generate the \code{latency}.
}
\examples{
\dontrun{
wa <- loadLatest() \%>\% monitor_subset(stateCodes = 'WA')
wa_current <- monitor_currentData(wa)
}
}
\keyword{ws_monitor}
