% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_dailyBarPlot.R
\name{monitor_dailyBarPlot}
\alias{monitor_dailyBarPlot}
\title{dailyBarPlot}
\usage{
monitor_dailyBarPlot(ws_monitor, monitorID = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}
}
\description{
A bar graph showing daily average PM 2.5 values for
a specific monitor. Each graph is colored with AQI Breaks.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150820)
monitor <- ws_monitor$meta$monitorID[1]
monitor_dailyBarPlot(ws_monitor, monitor)
}
}

