% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_dailyStatistic.R
\name{monitor_dailyStatistic}
\alias{monitor_dailyStatistic}
\title{Calculate Daily Statistics from Monitor Data}
\usage{
monitor_dailyStatistic(ws_monitor, FUN = get("mean"), dayStart = "midnight",
  na.rm = TRUE)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{FUN}{function used to collapse a day's worth of data into a single number}

\item{dayStart}{one of "sunset|midnight|sunrise"}

\item{na.rm}{logical value indicating whether NA values should be ignored}
}
\value{
Monitor data list with daily statistics.
}
\description{
Calculate daily statistics for each monitor in the \code{ws_monitor} data list.
}
\details{
Sunrise and sunset times are calculated based on the first monitor encountered.
This should be accurate enough for all use cases involving co-located monitors. Monitors
from different regions should have daily statistics calculated separately.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke')
airnow <- airnow_load(20150801, 20150831)
WA_smoky <- monitor_subset(airnow, stateCodes='WA', vlim=c(55,Inf))
WA_smoky_dailyMean <- monitor_dailyStatistic(WA_smoky, FUN=get('mean'), dayStart='midnight')
monitor_timeseriesPlot(WA_smoky_dailyMean, type='s')
}
}
\keyword{monitor}

