% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_dailyThreshold.R
\name{monitor_dailyThreshold}
\alias{monitor_dailyThreshold}
\title{Calculate Daily Fraction Over Threshold from Monitor Data}
\usage{
monitor_dailyThreshold(ws_monitor, threshold = "unhealthy",
  dayStart = "midnight", na.rm = TRUE)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{threshold}{AQI level name (e.g. 'unhealthy') or threshold value at or above which a measurement is counted}

\item{dayStart}{one of "sunset|midnight|sunrise"}

\item{na.rm}{logical value indicating whether NA values should be ignored}
}
\value{
Monitor data list with daily threshold statistics.
}
\description{
Calculate daily threshold statistics for each monitor in the \code{ws_monitor} data list.
}
\details{
Sunrise and sunset times are calculated based on the first monitor encountered.
This should be accurate enough for all use cases involving co-located monitors. Monitors
from different regions should have daily statistics calculated separately.
}
\examples{
\dontrun{
AirNow <- airnow_load(20150801, 20150930)

WA_IDS <- AirNow$meta$monitorID[AirNow$meta$stateCode == 'WA']
WA <- monitor_subset(AirNow, monitorIDs=WA_IDS)
monitor_leaflet(WA)

# Pull out specific locations
Twisp <- monitor_subset(WA, monitorIDs='530470009')
Winthrop <- monitor_subset(WA, monitorIDs='530470010')
Omak <- monitor_subset(WA, monitorIDs='530470013')

# Generate 
Twisp_daily <- monitor_dailyThreshold(Twisp, 89, dayStart='midnight')
Winthrop_daily <- monitor_dailyThreshold(Winthrop, 89, dayStart='midnight')
Omak_daily <- monitor_dailyThreshold(Omak, 89, dayStart='midnight')

# Plot
cols <- c(adjustcolor('orange',0.6),
          adjustcolor('red',0.6),
          adjustcolor('blue',0.6))
legend <- c('Twisp','Winthrop','Omak')
monitor_timeseriesPlot(Twisp_daily, pch=16, col=cols[1])
monitor_timeseriesPlot(Winthrop_daily, pch=16, col=cols[2], add=TRUE)
monitor_timeseriesPlot(Omak_daily, pch=16, col=cols[3], add=TRUE)
legend('topright', legend=legend, col=cols, pch=16)
title('Hours per day Above Threshold (PM 2.5 = 89)')
}
}
\keyword{monitor}

