% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_distance.R
\name{monitor_distance}
\alias{monitor_distance}
\title{Calculate Distances From ws_monitor Monitors to Location of Interest}
\usage{
monitor_distance(ws_monitor, lon, lat)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{lon}{longitude of the location of interest}

\item{lat}{latitude of the location of interest}
}
\value{
vector of of distances (km)
}
\description{
This function returns the distances (km) between monitoring sites and
a location of interest. These distances can be used to create a mask identifying
monitors within a certain radius of the location of interest.
}
\examples{
N_M <- Northwest_Megafires
# Walla Walla
WW_lon <- -118.330278
WW_lat <- 46.065
distance <- monitor_distance(N_M, WW_lon, WW_lat)
closestIndex <- which(distance == min(distance))
distance[closestIndex]
N_M$meta[closestIndex,]
}
\seealso{
distance
}
\keyword{ws_monitor}

