% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_esriMap.R
\name{monitor_esriMap}
\alias{monitor_esriMap}
\title{Create an ESRI Map of ws_monitor Object}
\usage{
monitor_esriMap(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, width = 640, height = 640, centerLon = NULL,
  centerLat = NULL, zoom = NULL, maptype = "worldStreetMap",
  grayscale = FALSE, mapRaster = NULL, cex = par("cex") * 2,
  pch = 16, ...)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}

\item{slice}{either a time index or a function used to collapse the time
axis -- defautls to \code{get('max')}}

\item{breaks}{set of breaks used to assign colors}

\item{colors}{a set of colors for different levels of air quality data
determined by \code{breaks}}

\item{width}{width of image, in pixels}

\item{height}{height of image, in pixels}

\item{centerLon}{map center longitude}

\item{centerLat}{map center latitude}

\item{zoom}{map zoom level}

\item{maptype}{map type}

\item{grayscale}{logical, if TRUE the colored map tile is rendered into a
black & white image}

\item{mapRaster}{optional RGB Raster* object returned from
\code{esriMap_getMap})}

\item{cex}{character expansion for points}

\item{pch}{plotting character for points}

\item{...}{arguments passed on to \code{esriMap_plotOnStaticMap}
(\emph{e.g.} destfile, cex, pch, etc.)}
}
\value{
Plots a map loaded from arcGIS REST with points for each monitor
}
\description{
Creates an ESRI map of a \emph{ws_monitor} object.

#' Available \code{maptypes} include:
\itemize{
\item{natGeo}
\item{worldStreetMap}
\item{worldTopoMap}
\item{satellite}
\item{deLorme}
}

Additional base maps are found at:
\url{http://resources.arcgis.com/en/help/arcgis-rest-api/index.html#/Basemaps/02r3000001mt000000/}

If \code{centerLon}, \code{centerMap} or \code{zoom} are not specified,
appropriate values will be calcualted using data from the
\code{ws_monitor$meta} dataframe.
}
\examples{
\dontrun{
N_M <- Northwest_Megafires
# monitorLeaflet(N_M) # to identify Spokane monitorIDs
Spokane <- monitor_subsetBy(N_M, stringr::str_detect(N_M$meta$monitorID,'^53063'))
Spokane <- monitor_subset(Spokane, tlim=c(20150815, 20150831))
monitor_esriMap(Spokane)
}
}
\seealso{
\code{\link{esriMap_plotOnStaticMap}}
}
\keyword{ws_monitor}
