% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_isMonitor.R
\name{monitor_isMonitor}
\alias{monitor_isMonitor}
\title{Test for an correct structure of ws_monitor Object}
\usage{
monitor_isMonitor(ws_monitor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
\code{TRUE} if \code{ws_monitor} has the correct structure, \code{FALSE} otherwise.
}
\description{
The \code{ws_monitor} is checked for the 'ws_monitor' class name
and presence of core metadata columns:
\itemize{
  \item{monitorID -- per deployment unique ID}
  \item{longitude -- decimal degrees E}
  \item{latitude -- decimal degrees N}
  \item{elevation -- height above sea level in meters}
  \item{timezone -- olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}
}
\examples{
monitor_isEmpty(Carmel_Valley)
}
\keyword{ws_monitor}
