% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_leaflet.R
\name{monitor_leaflet}
\alias{monitor_leaflet}
\title{Leaflet Interactive Map of Monitoring Stations}
\usage{
monitor_leaflet(ws_monitor, slice = get("max"), breaks = 7,
  paletteFunc = colorRampPalette(c("gray90", "antiquewhite", "yellow", "red",
  "firebrick")), radius = 10, opacity = 0.7, providerTiles = NULL,
  popupInfo = c("siteName", "monitorID", "elevation"))
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{slice}{either a time index or a function used to collapse the time axis -- defautls to \code{get('max')}}

\item{breaks}{set of breaks used to assign colors or a single integer used to provide quantile based breaks}

\item{paletteFunc}{a palette generating function as returned by \code{colorRampPalette}}

\item{radius}{radius of monitor circles}

\item{opacity}{opacity of monitor circles}

\item{providerTiles}{optional name of leaflet ProviderTiles to use, e.g. "Stamen.Terrain"}

\item{popupInfo}{a vector of column names from ws_monitor$meta to be shown in a popup window}
}
\value{
Initiates the interactive dygraph plot in Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in RStudio's 'Viewer' tab.  
Individual monitor timeseries are reduced to 
a single value by applying \code{statistic}to the entire timeseries of each monitor with \code{na.rm=TRUE}.
These values are then plotted over a map of the United States.

If \code{slice} is a function (not a function name) it will be used with argument \code{na.rm=TRUE} to
collapse the time dimension. Thus, user defined functions must accept \code{na.rm} as a parameter.
}
\details{
Using a single number for the \code{breaks} argument will cause the algorithm to use
quantiles to determine breaks.

You can use AQI colors and 24-hr, daily average breaks by specifying \code{breaks=NULL, paletteFunc=Null}.

See \url{https://leaflet-extras.github.io/leaflet-providers/} for a list of "provider tiles"
to use as the background map..
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
v_low <- AQI$breaks_24[4]
CA_unhealthy_monitors <- monitor_subset(airnow, stateCodes='CA', vlim=c(v_low, Inf))
monitor_leaflet(CA_unhealthy_monitors, providerTiles="Stamen.Terrain")
}
}
\keyword{ws_monitor}

