% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_map.R
\name{monitor_map}
\alias{monitor_map}
\title{Create Map of Monitoring Stations}
\usage{
monitor_map(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, showCounties = TRUE, showLegend = TRUE,
  add = FALSE, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{slice}{either a time index or a function used to collapse the time axis}

\item{breaks}{set of breaks used to assign colors}

\item{colors}{set of colors must be one less than the number of breaks}

\item{showCounties}{logical specifying whether to plot county boundaries -- defaults to \code{FALSE}}

\item{showLegend}{logical specifying whether to add a legend -- defaults to \code{TRUE}}

\item{add}{logical specifying whether to add points to an existing map}

\item{...}{additional arguments passed to points() such as 'col' or 'pch'}
}
\description{
Creates a map of monitoring stations in a given ws_monitor object. 
Individual monitor timeseries are reduced to a single value by applying the function 
passed in as \code{slice} to the entire timeseries of each monitor with \code{na.rm=TRUE}.
These values are then plotted over a map of the United States. Any additional 
arguments specified in '...' are passed on to the points() function.

If \code{slice} is an integer, it will be used as an index to pull out a single timestep.

If \code{slice} is a function (not a function name) it will be used with argument 
\code{na.rm=TRUE} to collapse the time dimension. Thus, any user defined functions
passed in as \code{slice} must accept \code{na.rm} as a parameter.
}
\details{
Using a single number for the \code{breaks} argument will result in the use
of quantiles to determine a set of breaks appropriate for the number of colors.
}
\examples{
\dontrun{
airnow <- airnow_load(20150901, 20150930)
nw <- monitor_subset(airnow, stateCodes=c('WA','OR'))
airnow_conus_daily <- monitor_dailyStatistic(airnow_conus, FUN=mean)
monitor_map(airnow_conus_daily)
title('Max Daily PM2.5 Levels in September, 2015')
}
}
\keyword{ws_monitor}

