% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_nowcast.R
\name{monitor_nowcast}
\alias{monitor_nowcast}
\title{Apply Nowcast Algorithm to ws_monitor Object}
\usage{
monitor_nowcast(ws_monitor, version = "pm")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{version}{character identify specifying the type of nowcast algorithm to be used}
}
\value{
ws_monitor object with data that have been proccessed by the Nowcast algorithm.
}
\description{
A Nowcast algorithm is applied to the data in in the ws_monitor object. The 
\code{version} argument specifies the minimum weight factor and number of hours to be 
considered in the calculation.

Available versions include:
\enumerate{
\item{pm}
\item{pmAsian}
\item{ozone}
}
}
\examples{
\dontrun{
ozone <- monitor_nowcast(ozone, 'ozone')
}
}
\references{
\url{https://en.wikipedia.org/wiki/Nowcast_(Air_Quality_Index)}

\url{https://www3.epa.gov/airnow/ani/pm25_aqi_reporting_nowcast_overview.pdf}

\url{http://aqicn.org/faq/2015-03-15/air-quality-nowcast-a-beginners-guide/}
}
\keyword{nowcast}

