% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_performanceMap.R
\name{monitor_performanceMap}
\alias{monitor_performanceMap}
\title{Create Map of Monitor Prediction Performance}
\usage{
monitor_performanceMap(predicted, observed, threshold = AQI$breaks_24[3],
  sizeBy = NULL, colorBy = "heidikeSkill",
  paletteFunc = grDevices::colorRampPalette(RColorBrewer::brewer.pal(6,
  "Purples")[-1]), breaks = c(-Inf, 0.5, 0.6, 0.7, 0.8, 0.9), ...)
}
\arguments{
\item{predicted}{data list of class \code{ws_monitor}}

\item{observed}{data list of class \code{ws_monitor}}

\item{threshold}{value used to classify \code{predicted} and \code{observed} measurements}

\item{sizeBy}{name of the metric used to create relative sizing}

\item{colorBy}{name of the metric used to create relative colors}

\item{paletteFunc}{a palette generating function as returned by \code{colorRampPalette}}

\item{breaks}{set of breaks used to assign colors or a single integer used to provide quantile based breaks - Must also specify the colorBy paramater}

\item{...}{additional arguments to be passed to the plot() funciton such as graphical parameters (see par)}
}
\value{
Dataframe of monitors vs named measure of performance.
}
\description{
This function uses "confusion matrix" analysis to calculate
different measures of predictive performance for every timeseries found
in \code{predicted} with respect to the observed values found in the single timeseries
found in \code{observed}.

Using a single number for the \code{breaks} argument will cause the algorithm to use
quantiles to determine breaks.
}
\details{
Setting either \code{sizeBy} or \code{colorBy} to \code{NULL} will cause
the size/colors to remain constant.
}
\examples{
\dontrun{
# Spokane summer of 2015
airnow <- airnow_load(20150701,20150930)
airnow <- monitor_rollingMean(airnow, width=3)
WA <- monitor_subset(airnow, stateCode='WA')
MonroeSt <- monitor_subset(WA, monitorIDs=530630047)
monitor_performanceMap(WA, MonroeSt, cex=2)
title('Heidike Skill of monitors predicting another monitor.')
}
}
\seealso{
\link{monitor_performance}
}
\keyword{monitor}

