% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_reorder.R
\name{monitor_reorder}
\alias{monitor_reorder}
\title{Reorder a ws_monitor Object}
\usage{
monitor_reorder(ws_monitor, monitorIDs = NULL)
}
\arguments{
\item{ws_monitor}{emph{ws_monitor} object}

\item{monitorIDs}{optional vector of monitor IDs used to reorder the meta and data dataframes}
}
\value{
A emph{ws_monitor} object reordered to match \code{monitorIDs}.
}
\description{
This function is a convenience function that merely wraps the \link{monitor_subset} function
which reorders as well as subsets.
}
\keyword{ws_monitor}

