% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_rollingMean.R
\name{monitor_rollingMean}
\alias{monitor_rollingMean}
\title{Compute Rolling Means of Monitor Data}
\usage{
monitor_rollingMean(ws_monitor, width = 8, data.thresh = 75,
  align = "center")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{width}{width of window (in hours) used to calculate means}

\item{data.thresh}{minimum number of observations as a percent of width required, or NA is returned}

\item{align}{alignment of averaging window relative to point being calculated ["left"|"center"|"right"]}
}
\value{
\code{ws_monitor} object with rolling mean data
}
\description{
Calculate rolling means for each monitor in the \code{ws_monitor} object using the
\code{openair::rollingMean()} function
}
\examples{
\dontrun{
airnow <- airnow_load(20150801, 20150808)
WA_smoky <- monitor_subset(airnow, stateCodes='WA', vlim=c(100,Inf))
WA_smoky_3hr <- monitor_rollingMean(WA_smoky, width=3, align="center")
WA_smoky_24hr <- monitor_rollingMean(WA_smoky, width=24, align="right")
monitor_timeseriesPlot(WA_smoky, type='l', shadedNight=TRUE)
monitor_timeseriesPlot(WA_smoky_3hr, type='l', col='red', add=TRUE)
monitor_timeseriesPlot(WA_smoky_24hr, type='l', col='blue', lwd=2, add=TRUE)
legend('topright', c("hourly","3-hourly","24-hourly"),
       col=c('black','red','blue'), lwd=c(1,1,2))
title('Smoky Monitors in Washington -- August, 2015')
}
}
\keyword{monitor}

