% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_rollingMeanPlot.R
\name{monitor_rollingMeanPlot}
\alias{monitor_rollingMeanPlot}
\title{dailyBarPlot}
\usage{
monitor_rollingMeanPlot(ws_monitor, monitorID = NULL, width = 24,
  data.thresh = 75, align = "center")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}

\item{width}{width of window (in hours) used to calculate means}

\item{data.thresh}{minimum number of observations as a percent of width required, or NA is returned}

\item{align}{alignment of averaging window relative to point being calculated ["left"|"center"|"right"]}
}
\description{
A bar graph showing daily average PM 2.5 values for
a specific monitor. Each graph is colored with AQI Breaks.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150725, 20150805)
monitor <- ws_monitor$meta$monitorID[1]
monitor_rollingMeanPlot(ws_monitor, monitor)
}
}

