% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_scaleData.R
\name{monitor_scaleData}
\alias{monitor_scaleData}
\title{Scale Monitoring Data}
\usage{
monitor_scaleData(ws_monitor, expression)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}

\item{expression}{an \R{R} expression used to scale data}
}
\description{
Use an R expression scale the data in a \emph{ws_monitor} object.

The \R{R} expression given in \code{expression} is applied to all data columns in
in \code{ws_monitor$data}.

Expressions should use \code{data} for the left hand side of the expression.
}
\examples{
wa <- monitor_subset(Northwest_Megafires, stateCodes='WA')
wa_zero <- monitor_scaleData(wa, data * 3.4)
}
\keyword{ws_monitor}

