% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_subset.R
\name{monitor_subset}
\alias{monitor_subset}
\title{Subset Monitoring Data}
\usage{
monitor_subset(ws_monitor, xlim = NULL, ylim = NULL, tlim = NULL,
  vlim = NULL, monitorIDs = NULL, stateCodes = NULL,
  dropMonitors = TRUE)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{xlim}{optional longitude lim with lo and hi longitude values}

\item{ylim}{optional latitude lim with lo and hi latitude values}

\item{tlim}{optional time lim with lo and hi time values (POSIXct)}

\item{vlim}{optional data lim with lo and hi data values}

\item{monitorIDs}{optional vector of monitorIDs}

\item{stateCodes}{optional vector of stateCodes}

\item{dropMonitors}{flag specifying whether to remove monitors with no data}
}
\value{
monitoring data list filtered as specified
}
\description{
The incoming monitoring data list is filtered according to the parameters
passed in.  If any parameter is not specified, that parameter will not be used in the filtering.
}
\details{
By default, filtering by tlim or vlim will always return a 'data' dataframe with the
    same number of columns as the incoming dataframe. Some columns may consist sof all \code{NA}s.
    If \code{dropMonitors=TRUE}, columns
    will be removed if there are not valid data for a specific monitor after subsetting.
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
xlim <- c(-124.801, -122.801)
ylim <- c(47.004, 48.404)
Olympic_Peninsula <- monitor_subset(airnow, xlim, ylim)
monitor_map(Olympic_Peninsula)
}
}
\keyword{ws_monitor}

