% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_subsetByDistance.R
\name{monitor_subsetByDistance}
\alias{monitor_subsetByDistance}
\title{Subset ws_monitor Object by Distance from Target Location}
\usage{
monitor_subsetByDistance(ws_monitor, lon = NULL, lat = NULL, radius = 50,
  count = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{lon}{target longitude from which the radius will be calculated}

\item{lat}{target latitude from which the radius will be calculated}

\item{radius}{distance (km) of radius from target location -- default=300}

\item{count}{number of grid cells to return}
}
\value{
\code{ws_monitor} object containing monitors near a location
}
\description{
Subsets a ws_monitor object to include only those monitors (or grid cells)
within a certain radius of a target location. If no monitors (or grid cells) fall
within the specified \code{radius}, \code{ws_monitor$data} and \code{ws_monitor$meta}
are set to \code{NULL}.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke')
airnow <- airnow_load(20140913, 20141010)
King_Fire <- monitor_subsetByDistance(airnow, lon=-120.604, lat=38.782, radius=50)
monitor_leaflet(King_Fire)
}
}
\seealso{
monitorDistance
}
\keyword{ws_monitor}

