% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_subsetMeta.R
\name{monitor_subsetMeta}
\alias{monitor_subsetMeta}
\title{Subset ws_monitor Object 'meta' Dataframe}
\usage{
monitor_subsetMeta(
  meta,
  xlim = NULL,
  ylim = NULL,
  stateCodes = NULL,
  countryCodes = NULL,
  monitorIDs = NULL
)
}
\arguments{
\item{meta}{\emph{ws_monitor} object \code{meta} dataframe}

\item{xlim}{optional vector with low and high longitude limits}

\item{ylim}{optional vector with low and high latitude limits}

\item{stateCodes}{optional vector of stateCodes}

\item{countryCodes}{optional vector of countryCodes}

\item{monitorIDs}{optional vector of monitorIDs}
}
\value{
A \emph{ws_monitor} object \code{meta} dataframe, or \code{NULL} if filtering removes all monitors.
}
\description{
Subsets the \code{ws_monitor$data} dataframe by removing any monitors that
lie outisde the geographical ranges specified (i.e. outside of the given longitudes and 
latitudes and/or states) and that are not mentioned in the list of monitorIDs.

If any parameter is not specified, that parameter will not be used in the subsetting.

Intended for use by the monitor_subset function.
}
\details{
Longitudes must be specified in the domain [-180,180].
}
\keyword{ws_monitor}
