% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_timeAverage.R
\name{monitor_timeAverage}
\alias{monitor_timeAverage}
\title{Calculate Time Averages}
\usage{
monitor_timeAverage(ws_monitor, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{...}{additional arguments to be passed to openair::timeAverage()}
}
\value{
ws_monitor object with data that have been proccessed by \code{openair::timeAverage()}
}
\description{
This function extracts the \code{data} dataframe from the \code{ws_monitor}
and renames the 'datetime' column so that it can be processed by the \code{openair} package's
\code{timeAverage()} function. (See that function for details.)
}
\examples{
\dontrun{
airnow <- airnow_load(20160801,20160831)
CarmelValley <- monitor_subset(airnow, monitorIDs="060530002",
                               tlim=c(2016080800,2016081023),
                               timezone='America/Los_Angeles')
CarmelValley_3hrChunks <- monitor_timeAverage(CarmelValley, avg.time="3 hour")
layout(matrix(seq(2)))
monitorPlot_hourlyBarplot(CarmelValley, main='1-Hourly Average PM2.5')
monitorPlot_hourlyBarplot(CarmelValley_3hrChunks, main='3-Hourly Average PM2.5')
layout(1)
}
}
\keyword{nowcast}

