% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_timeOfDayMultiPiePlot.R
\name{monitor_timeOfDayMultiPiePlot}
\alias{monitor_timeOfDayMultiPiePlot}
\title{Create 9 Daily Time of Day Pie Plots}
\usage{
monitor_timeOfDayMultiPiePlot(ws_monitor, monitorID = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor object (optional if only one monitor
in the ws_monitor object)}
}
\description{
Creates pie plots of 3-hour average PM2.5 concentrations over 9 consecutive days for a single monitor.
3-hour pie pieces are colored according to AQI category. Day/night is approximated using shading.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150810)
monitor <- ws_monitor$meta$monitorID[1]
monitor_timeOfDayMultiPiePlot(ws_monitor, monitor)
}
}
\keyword{ws_monitor}

