% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_timeOfDaySpaghettiPlot.R
\name{monitor_timeOfDaySpaghettiPlot}
\alias{monitor_timeOfDaySpaghettiPlot}
\title{Create Time of Day Spaghetti Plot}
\usage{
monitor_timeOfDaySpaghettiPlot(ws_monitor, monitorID = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor object}
}
\description{
Creates a spaghetti plot of PM2.5 levels by hour for one or more days. The average by hour over
the period is also calculated and plotted as a thick black line.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150810)
monitor <- ws_monitor$meta$monitorID[1]
monitor_timeOfDaySpaghettiPlot(ws_monitor, monitor)
}
}
\keyword{ws_monitor}

