% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_trim.R
\name{monitor_trim}
\alias{monitor_trim}
\title{Trim a Monitor Time Axis}
\usage{
monitor_trim(ws_monitor)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}
}
\value{
A \code{ws_monitor} objects.
}
\description{
The incoming \code{ws_monitor} object has its time axis trimmed
to include everything from the first to last valid datapoints for any monitor
in the object. All timesteps before or after any valid data are removed
}
\keyword{monitor}

