% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorGoogleMap.R
\name{monitorGoogleMap}
\alias{monitorGoogleMap}
\title{Create a Google Map of ws_monitor Object}
\usage{
monitorGoogleMap(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, labels = AQI$names, legendTitle = "Max AQI Level",
  legendX = "topright", legendY = NULL, showLegend = TRUE, width = 640,
  height = 640, zoom = NULL, maptype = "terain", grayscale = FALSE,
  map = NULL, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{slice}{either a time index or a function used to collapse the time axis -- defautls to \code{get('max')}}

\item{breaks}{set of breaks used to assign colors}

\item{colors}{a set of colors for different levels of air quality data determined by \code{breaks}}

\item{labels}{a set of text labels, one for each color}

\item{legendTitle}{legend title}

\item{legendX}{x coordinate passed on to the legend() command}

\item{legendY}{y coordinate passed on to the legend() command}

\item{showLegend}{logical specifying whether to add a legend}

\item{width}{width of image, in pixels}

\item{height}{height of image, in pixels}

\item{zoom}{map zoom level}

\item{maptype}{map type}

\item{grayscale}{logical, if TRUE the colored map tile is rendered into a black & white image}

\item{map}{optional map object returned from monitorGoogleMap()}

\item{...}{arguments passed on to RgoogleMaps::PlotOnStaticMap() (e.g. destfile, cex, pch, etc.)}
}
\value{
A \code{MyMap} RgoogleMaps map object object that can serve as a base plot.
}
\description{
Creates a Google map of a ws_monitor object.
TODO:  More description here.
}
\examples{
\dontrun{
CarmelValley <- airnow_load(20160801,20160831,monitorIDs="060530002")
monitorGoogleMap(CarmelValley)
}
}
\keyword{ws_monitor}

