% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorMap_performance.R
\name{monitorMap_performance}
\alias{monitorMap_performance}
\title{Create Map of Monitor Prediction Performance}
\usage{
monitorMap_performance(predicted, observed, threshold = AQI$breaks_24[3],
  cex = par("cex"), sizeBy = NULL, colorBy = "heidikeSkill",
  breaks = c(-Inf, 0.5, 0.6, 0.7, 0.8, Inf),
  paletteFunc = grDevices::colorRampPalette(RColorBrewer::brewer.pal(length(breaks),
  "Purples")[-1]), showLegend = TRUE, stateCol = "grey60", stateLwd = 2,
  countyCol = "grey70", countyLwd = 1, add = FALSE, ...)
}
\arguments{
\item{predicted}{ws_monitor object with predicted values}

\item{observed}{ws_monitor object with observed  values}

\item{threshold}{value used to classify \code{predicted} and \code{observed} measurements}

\item{cex}{the amount that the points will be magnified on the map}

\item{sizeBy}{name of the metric used to create relative sizing}

\item{colorBy}{name of the metric used to create relative colors}

\item{breaks}{set of breaks used to assign colors or a single integer used to provide quantile based breaks - Must also specify the colorBy paramater}

\item{paletteFunc}{a palette generating function as returned by \code{colorRampPalette}}

\item{showLegend}{logical specifying whether to add a legend (default: \code{TRUE})}

\item{stateCol}{color for state outlines on the map}

\item{stateLwd}{width for state outlines}

\item{countyCol}{color for county outline on the map}

\item{countyLwd}{width for county outlines}

\item{add}{logical specifying whether to add to the current plot}

\item{...}{additional arguments to be passed to the \code{maps::map()} funciton such as graphical parameters (see code{?par})}
}
\description{
This function uses \emph{confusion matrix} analysis to calculate
different measures of predictive performance for every timeseries found
in \code{predicted} with respect to the observed values found in the single timeseries
found in \code{observed}.

Using a single number for the \code{breaks} argument will cause the algorithm to use
quantiles to determine breaks.
}
\details{
Setting either \code{sizeBy} or \code{colorBy} to \code{NULL} will cause
the size/colors to remain constant.
}
\examples{
\dontrun{
# Spokane summer of 2015
wa <- airnow_load(20150701, 20150930, stateCodes='WA')
wa <- monitor_rollingMean(wa, width=3)
MonroeSt <- monitor_subset(wa, monitorIDs=530630047)
monitorMap_performance(wa, MonroeSt, cex=2)
title('Heidike Skill of monitors predicting another monitor.')
}
}
\seealso{
\link{monitor_performance}
}
\keyword{ws_monitor}

