% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_dailyBarplotFacets.R
\name{monitorPlot_dailyBarplotFacets}
\alias{monitorPlot_dailyBarplotFacets}
\title{Create Daily AQI Barplots of Multiple Monitors}
\usage{
monitorPlot_dailyBarplotFacets(ws_monitor, monitorIDs, title,
  smooth_func = monitor_nowcast, smooth_args = list(),
  smooth_style = c("bars", "points"), smooth_name = "Hourly NowCast",
  theme = theme_monitors(), ...)
}
\arguments{
\item{ws_monitor}{\code{ws_monitor} object}

\item{monitorIDs}{vector of monitor IDs in ws_monitor to subset and plot}

\item{title}{plot title}

\item{smooth_func}{PWFSLSmoke smoothing function}

\item{smooth_args}{list of additional arguments to the smoothing function}

\item{smooth_style}{how smoothed values are plotted (set to \code{NULL} to exclude smoothed values)}

\item{smooth_name}{name of smoothing method used in legend}

\item{theme}{\code{ggplot2::theme} controlling plotting options}

\item{...}{additional arguments, such as \code{tlim}, passed to \code{monitor_subset()}}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
Plots a facetted set of air quality monitoring data.
}
\keyword{ws_monitor}
