% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_hourlyBarplot.R
\name{monitorPlot_hourlyBarplot}
\alias{monitorPlot_hourlyBarplot}
\title{Create Hourly Barplot}
\usage{
monitorPlot_hourlyBarplot(ws_monitor, monitorID = NULL, tlim = NULL,
  localTime = TRUE, style = "AQI", shadedNight = TRUE, gridPos = "",
  gridCol = "black", gridLwd = 0.5, gridLty = "solid", dayCol = "black",
  dayLwd = 2, dayLty = "solid", hourCol = "black", hourLwd = 1,
  hourLty = "solid", hourInterval = 6, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the given ws_monitor object (optional
if only one monitor in the ws_monitor object)}

\item{tlim}{optional vector with start and end times (integer or character representing YYYYMMDD[HH])}

\item{localTime}{logical specifying whether \code{tlim} is in local time or UTC}

\item{style}{named style specification ('AirFire')}

\item{shadedNight}{add nighttime shading}

\item{gridPos}{position of grid lines either 'over', 'under' ('' for no grid lines)}

\item{gridCol}{grid color}

\item{gridLwd}{grid line width}

\item{gridLty}{grid line type}

\item{dayCol}{day boundary color}

\item{dayLwd}{day boundary line width (set to 0 to omit day lines)}

\item{dayLty}{day boundary type}

\item{hourCol}{hour boundary color}

\item{hourLwd}{hour boundary line width (set to 0 to omit  hour lines)}

\item{hourLty}{hour boundary type}

\item{hourInterval}{interval for hour boundary lines}

\item{...}{additional arguments to be passed to barplot()}
}
\description{
Creates a bar plot showing hourly PM 2.5 values for a specific monitor in a ws_monitor object.
Colors are assigned to one of the following styles:
\itemize{
\item{\code{AQI} -- hourly values colored with AQI colors using AQI 24-hour breaks}
\item{\code{brownScaleAQI} -- hourly values colored with brownscale colors using AQI 24-hour breaks}
\item{\code{grayScaleAQI} -- hourly values colored grayscale colors using AQI 24-hour breaks}
}
}
\examples{
\dontrun{
airnow <- airnow_load(20150701, 20150710)
title <- "Hourly PM2.5 for Colleville, WA"
monitorPlot_hourlyBarplot(airnow, monitorID="530650004")
}
}
\keyword{ws_monitor}

