% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_locationMap.R
\name{monitorPlot_locationMap}
\alias{monitorPlot_locationMap}
\title{Create Map of ws_monitor Object and Surrounding Events}
\usage{
monitorPlot_locationMap(ws_monitor, monitorIcon = NULL,
  eventLocations = NULL, eventIcon = NULL, width = 250, height = 250,
  zoom = 9, maptype = "terain")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorIcon}{optional icon to use in lieu of default blue paddle}

\item{eventLocations}{optional dataframe containing longitude and latitude data for events (e.g. fires)}

\item{eventIcon}{icon for optional events}

\item{width}{width of image, in pixels}

\item{height}{height of image, in pixels}

\item{zoom}{map zoom level}

\item{maptype}{map type}
}
\description{
Creates a map of a ws_monitor object with icons indicating monitor(s) and surrounding events, if 
specified.
}
\examples{
\dontrun{
ws_monitor <- airsis_load(20150901, 20150930)
monitor <- ws_monitor$meta$monitorID[3]
monitorPlot_timeOfDaySpaghetti(ws_monitor, monitor,tlim=c(20150914,20150930))
}
}
\keyword{ws_monitor}

