% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_noData.R
\name{monitorPlot_noData}
\alias{monitorPlot_noData}
\title{Create Blank Plot with "No Data" Message}
\usage{
monitorPlot_noData(ws_monitor, monitorID = NULL, cex = 2.5)
}
\arguments{
\item{ws_monitor}{emph{ws_monitor} object}

\item{monitorID}{id for a specific monitor in the given ws_monitor object}

\item{cex}{text size of "No data for"; monitor ID is one size smaller}
}
\description{
Creates a blank plot that indicates there is no data for the monitor of interest
}
\keyword{internal}

