% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_timeOfDaySpaghetti.R
\name{monitorPlot_timeOfDaySpaghetti}
\alias{monitorPlot_timeOfDaySpaghetti}
\title{Create Time of Day Spaghetti Plot}
\usage{
monitorPlot_timeOfDaySpaghetti(ws_monitor, monitorID = NULL, tlim = NULL,
  ylim = NULL, aqiLines = TRUE, shadedNight = TRUE, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor object}

\item{tlim}{time limit for plot}

\item{ylim}{optional y limits for plot}

\item{aqiLines}{horizontal lines indicating AQI levels}

\item{shadedNight}{shade nights based on sunrise/sunset times of middle day in selected period}

\item{...}{additional arguments to pass to lines()}
}
\description{
Creates a spaghetti plot of PM2.5 levels by hour for one or more days. The average by hour over 
the period is also calculated and plotted as a thick red line.
}
\examples{
\dontrun{
ws_monitor <- airsis_load(20150901, 20150930)
monitor <- ws_monitor$meta$monitorID[3]
monitorPlot_timeOfDaySpaghetti(ws_monitor, monitor,tlim=c(20150914,20150930))
}
}
\keyword{ws_monitor}

