% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{Northwest_Megafires}
\alias{Northwest_Megafires}
\title{Northwest_Megafires example dataset}
\format{
A \emph{ws_monitor} object with "meta" and "data" dataframes.
}
\usage{
Northwest_Megafires
}
\description{
In the summer of 2015 Washington state had several catastrophic wildfires that led
to many days of heavy smoke in eastern Washington, Oregon and northern Idaho.
The Northwest_Megafires dataset contains AirNow ambient monitoring data for the
Pacific Northwest from May 31 through November 01, 2015 (UTC).

The \code{Northwest_Megafires} dataset provides a quickly loadable
version of a \emph{ws_monitor} object for practicing and code examples.
}
\keyword{datasets}
