% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq_assignStateCode.R
\name{openaq_assignStateCode}
\alias{openaq_assignStateCode}
\title{Assign State Codes to the Available Longitudes and Latitudes}
\usage{
openaq_assignStateCode(df)
}
\arguments{
\item{df}{a dataframe with longitude and latitude columns}
}
\value{
Input dataframe with additional columns: \code{stateCode}.
}
\description{
Uses function getStateCode() from MazamaSpatialUtils package 
to assign state codes to the longitudes and latitudes in the dataframe.
}
\examples{
\dontrun{
df <- openaq_downloadData(startdate = 20160901, days=2)
df <- openaq_assignStateCode(df)
}
}
\keyword{OpenAQ}

