% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openaq_createDataDataFrame.R
\name{openaq_createDataDataframe}
\alias{openaq_createDataDataframe}
\title{Create Data Dataframe}
\usage{
openaq_createDataDataframe(df)
}
\arguments{
\item{df}{a OpenAQ dataframe after metadata enhancement}
}
\value{
A \code{'data'} dataframe for use in a \code{ws_monitor} object.
}
\description{
After addtional columns(i.e. \code{'datetime'}, \code{'stateCode'}, and \code{'monitorID'})
have been applied to an OpenAQ dataframe,
we can extract the PM2.5 values and store them in a \code{'data'} dataframe
organized as hour-by-monitor.

The first column of the returned dataframe is named \code{'datetime'} and
contains a \code{POSIXct} time in UTC. Additional columns contain data
for each separate monitorID.
}
\keyword{OpenAQ}

