% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openaq_createMetaDataFrame.R
\name{openaq_createMetaDataframe}
\alias{openaq_createMetaDataframe}
\title{Create Sites Metadata Dataframe}
\usage{
openaq_createMetaDataframe(df)
}
\arguments{
\item{df}{a OpenAQ dataframe after metadata enhancement}
}
\value{
A \code{'meta'} dataframe for use in a \code{ws_monitor} object.
}
\description{
After an OpenAQ dataframe has been enhanced with
additional columns including \code{'datetime'}, \code{'stateCode'}, \code{'monitorID'} we are ready to
pull out site information associated with unique monitorID.

These will be rearranged into a dataframe organized as site-by-property
with one row for each monitorID.

This site information found in \code{df} is augmented so that we end up with a uniform
set of properties associated with each monitorID. The list of
columns in the returned \code{'meta'} dataframe is:

\preformatted{
> names(meta)
[1] "siteName"    "latitude"    "longitude"   "timezone"    "countryCode" "stateCode"   "monitorID"
}
}
\keyword{OpenAQ}

