% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq_createMonitorObject.R
\name{openaq_createMonitorObject}
\alias{openaq_createMonitorObject}
\title{Obtain OpenAQ Data and Create ws_monitor Object}
\usage{
openaq_createMonitorObject(countryCode = "US", parameter = "pm25",
  startdate = "", days = 1, verbose = FALSE, saveFile = NULL)
}
\arguments{
\item{countryCode}{two-character country code (ISO 3166-1 alpha-2)}

\item{parameter}{pollutant name}

\item{startdate}{desired staring date (integer or character representing YYYYMMDD)}

\item{days}{desired number of days of data to assemble}

\item{verbose}{logical flag to generate verbose output}

\item{saveFile}{optional filename where raw CSV will be written}
}
\value{
ws_monitor object with a unique `monitorID`.
}
\description{
Obtains monitor data from OPENAQ and converts
it into a metadata enhanced \code{ws_monitor} object
ready for use with all \code{monitor_~} functions.

Available parameters are limited to:
PM2.5, PM10, ozone (O3), sulfur dioxide (SO2), nitrogen dioxide (NO2), 
carbon monoxide (CO), and black carbon (BC)
\enumerate{
\item{pm25}{ -- PM2.5}
\item{pm10}{ -- PM10}
\item{o3}{ -- ozone }
\item{so2}{ -- sulfer dioxide }
\item{no2}{ -- nitrogen dioxide}
\item{co}{ -- carbon monoxide}
\item{bc}{ -- black carbon}
}
}
\examples{
\dontrun{
df <- openaq_downloadData(20160901, days=7)
}
}
\keyword{OpenAQ}

