% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq_load.R
\name{openaq_load}
\alias{openaq_load}
\title{Load OpenAQ Monitoring Data}
\usage{
openaq_load(startdate, enddate, monitorIDs = NULL, parameter = "pm25",
  baseUrl = "http://smoke.airfire.org/RData/OpenAQ/")
}
\arguments{
\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{monitorIDs}{The set of monitor IDs to be subsetted.
When set to NULL, the function returns all the monitor IDs available in the metadata}

\item{parameter}{The parameter of interest}

\item{baseUrl}{base URL for OpenAQ meta and data files}
}
\value{
ws_monitor object with subsetted time, monitorIDs and parameter
}
\description{
When given the startdate, enddate, monitorIDs and parameter of interest, the function retrieves the 
meta and data files from the archive baseUrl (or local directory) and return the subsetted ws_monitor object.

#' Available parameters include:
\enumerate{
\item{pm25}{ -- PM2.5}
\item{pm10}{ -- PM10}
\item{o3}{ -- ozone }
\item{so2}{ -- sulfer dioxide }
\item{no2}{ -- nitrogen dioxide}
\item{co}{ -- carbon monoxide}
\item{bc}{ -- black carbon}
}
}
\examples{
\dontrun{
openaq <- openaq_load(20150901, 20150930)
openaq_conus <- monitor_subset(openaq, stateCodes=CONUS)
monitor_leaflet(openaq_conus)
}
}
\keyword{OpenAQ}

