% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_enhance.R
\name{raw_enhance}
\alias{raw_enhance}
\title{Process Raw Monitoring Data to Create raw_enhance Object}
\usage{
raw_enhance(raw, rawSource = "AIRSIS")
}
\arguments{
\item{raw}{raw monitor data, as created by airsis_createRawDataframe or wrcc_createRawDataframe}

\item{rawSource}{source of raw monitor data (i.e. "AIRSIS" or "WRCC")}
}
\value{
Dataframe with original raw data, plus new columns with raw naming scheme for downstream use.
}
\description{
Processes raw monitor data to create a consistent format that can be handled by various raw* functions.
All original raw data is retained, and additional columns are added with names add consistently-named columns enabling
processing from various sources.
}
\examples{
\dontrun{
raw <- airsis_createRawDataframe(startdate = 20160901,enddate=20161015, unitID = 1012)
raw <- raw_enhance(raw,rawSource='AIRSIS')
View(raw)
}
}
\keyword{raw_enhance}

