% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_timeOfDaySpaghettiPlot.R
\name{raw_timeOfDaySpaghettiPlot}
\alias{raw_timeOfDaySpaghettiPlot}
\title{Time of Day Spaghetti Plot}
\usage{
raw_timeOfDaySpaghettiPlot(df, dataVar, tzone = NULL, col = "salmon",
  meanCol = "black", meanLwd = 4, highlightDates = c(),
  highlightCol = "dodgerblue", ...)
}
\arguments{
\item{df}{dataframe with \code{datetime} column in GMT}

\item{dataVar}{variable to be plotted}

\item{tzone}{timezone where data were collected}

\item{meanCol}{color used for the mean line (use NA to omit the mean)}

\item{meanLwd}{line width used for th mean line}

\item{highlightDates}{dates to be highlighted in YYYYMMDD format}

\item{highlightCol}{color used for highlighted days}

\item{...}{additional graphical parameters are passed to the lines() function}
}
\description{
Spaghetti Plot that shows data by hour-of-day.
}
\examples{
\dontrun{
}
}

