% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawPlot_pollutionRose.R
\name{rawPlot_pollutionRose}
\alias{rawPlot_pollutionRose}
\title{Create a Pollution Rose Plot from raw_enhance Object}
\usage{
rawPlot_pollutionRose(df, parameter = "pm25", tlim = NULL,
  localTime = TRUE, normalize = FALSE, ...)
}
\arguments{
\item{df}{raw_enhance monitor object, as created by the raw_enhance() function}

\item{parameter}{parameter to plot}

\item{tlim}{A vector that subsets the raw dataframe by time limits, e.g. c(20160915,20161010)}

\item{localTime}{use local times; defaults to GMT if >1 time zone in data.}

\item{normalize}{normalize slices to fill entire area, allowing for easier comparison of counts of magnitudes by direction}

\item{...}{additional arguments to pass on to openair::pollutionRose()}
}
\description{
Create pollution rose plot from raw_enhance object based on openair::pollutionRose(). If normalized, black line
indicates frequency by direction.
}
\examples{
\dontrun{
df <- PWFSLSmoke::airsis_createRawDataframe(startdate = 20160901, enddate=20161015, unitID = 1012)
df <- raw_enhance(df,rawSource='AIRSIS')
rawPlot_pollutionRose(df)
}
}
\keyword{raw_enhance}

