% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawPlot_timeseries.R
\name{rawPlot_timeseries}
\alias{rawPlot_timeseries}
\title{Plot Raw Monitoring Data}
\usage{
rawPlot_timeseries(df, parameter = "pm25", localTime = TRUE, tlim = NULL,
  shadedNight = TRUE, shadedBackground = "windSpeed", type = NULL,
  sbLwd = 1, ...)
}
\arguments{
\item{df}{raw_enhance monitor object, as created by the raw_enhance() function}

\item{parameter}{raw parameter to plot. Default = "pm25". Other options include
c("temperature","humidity","windSpeed","windDir", "pressure"), or any of the other raw
parameters (do "names(df)" to see list of options)}

\item{localTime}{use local times; defaults to GMT if >1 time zone in data.}

\item{tlim}{A vector that subsets the raw dataframe by time limits, e.g. c(20160915,20161010)}

\item{shadedNight}{Shade background based on approximate sunrise/sunset times. Unavailable if >1 time zone in data.
Also note that for multiple deployments, this defaults to use the lat/lon for the first deployment, which in theory
could be somewhat unrepresentative, such as if deployments have a large range in latitude.}

\item{shadedBackground}{Add vertical lines corresponding to a second parameter; currently defaults to wind speed binned into quartiles. Future iterations
may include options to choose which parameter to plot, which color to use, which intervals, etc.}

\item{type}{Line type; defaults to best option for each type, unless specified otherwise in this argument. To replace w/ ...}

\item{sbLwd}{shaded background line width}

\item{...}{additional arguments to pass to lines() function}
}
\description{
Creates a plot of raw monitoring data as generated using raw_enhance().
}
\keyword{raw_enhance}

