% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawPlot_windRose.R
\name{rawPlot_windRose}
\alias{rawPlot_windRose}
\title{Create Wind Rose Plot from a Raw Dataframe}
\usage{
rawPlot_windRose(df, tlim = NULL, localTime = TRUE, ...)
}
\arguments{
\item{df}{enhanced, raw dataframe as created by the raw_enhance() function}

\item{tlim}{optional vector with start and end times (integer or character representing YYYYMMDD[HH])}

\item{localTime}{use local times; defaults to GMT if >1 time zone in data.}

\item{...}{additional arguments to pass on to openair::windRose()}
}
\description{
Create wind rose plot from raw_enhance object. Based on openair::windRose().
}
\examples{
\dontrun{
df <- PWFSLSmoke::airsis_createRawDataframe(startdate = 20160901,enddate=20161015, unitID = 1012)
df <- raw_enhance(df,rawSource='AIRSIS')
rawPlot_windRose(df)
}
}
\keyword{raw}

