% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{setEsriToken}
\alias{setEsriToken}
\title{Set ESRI Token}
\usage{
setEsriToken(token)
}
\arguments{
\item{token}{ESRI token used when interacting with ESRI location services}
}
\value{
Silently returns previous value of esriToken.
}
\description{
Sets the current esriToken.
}
\seealso{
addEsriAddress

getEsriToken

esriToken
}
\keyword{environment}
