% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{setGoogleApiKey}
\alias{setGoogleApiKey}
\title{Set Google API Key}
\usage{
setGoogleApiKey(key)
}
\arguments{
\item{key}{Google API key used when interacting with Google location services}
}
\value{
Silently returns previous value of googleApiKey.
}
\description{
Sets the current Google API key.
}
\seealso{
addGoogleAddress

addGoogleElevation

getGoogleApiKey

googleApiKey
}
\keyword{environment}
