% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{setSmokeDataDir}
\alias{setSmokeDataDir}
\title{Set Package Data Directory}
\usage{
setSmokeDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where smoke datasets are created}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where spatial data is located.
If the directory does not exist, it will be created.
}
\seealso{
\link{SmokeDataDir}

\link{getSmokeDataDir}
}
\keyword{environment}

