% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{SmokeDataDir}
\alias{SmokeDataDir}
\title{Directory for Smoke Data}
\format{Absolute path string.}
\description{
This package maintains an internal directory location which users can set
using \code{setSmokeDataDir()}. All package functions use this directory whenever datasets
are created or loaded.

The default setting when the package is loaded is \code{getwd()}.
}
\seealso{
\link{getSmokeDataDir}

\link{setSmokeDataDir}
}
\keyword{environment}

