% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgradeMeta_v1.0.R
\name{upgradeMeta_v1.0}
\alias{upgradeMeta_v1.0}
\title{Upgrade ws_monitor Metadata to Version 1.0}
\usage{
upgradeMeta_v1.0(ws_monitor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
A \emph{ws_monitor} object with version 1.0 metadata.
}
\description{
Upgrade a \emph{ws_monitor} object to version 1.0 standards.
}
