% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{US_52}
\alias{US_52}
\title{US State Codes}
\format{A vector with 52 elements}
\usage{
US_52
}
\description{
State codes for the 50 states +DC +PR (Puerto Rico)
}
\details{
US state codes
}
\keyword{datasets}
