% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{WRCC}
\alias{WRCC}
\title{WRCC Monitor Names and Unit IDs}
\format{A list of lists}
\description{
The WRCC \url{http://www.wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitor Archive}
provides access to a variety of monitors that can be accessed with the \link{wrcc_createMonitorObject}
function. Use of this funciton requires a valid unitID. The \code{WRCC} object is
a list of lists. The element named \code{unitIDs} is itself a list of three named vectors,
each containing the unitIDs and associated names for
one of the categories of monitors available at WRCC:

\itemize{
\item{cache}
\item{miscellaneous}
\item{usfs_regional}
}
}
\details{
WRCC monitor names and unitIDs
}
\note{
This list of monitor types was created on Feb 09, 2017.
}
\keyword{datasets}
