% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_buildLibrary.R
\name{wrcc_buildLibrary}
\alias{wrcc_buildLibrary}
\title{Build a Local RData Library from Local WRCC Data}
\usage{
wrcc_buildLibrary(dataDir = "~/Data/WRCC", outputDir = "~/Data/WRCC",
  logLevel = "INFO", transcript = "WRCC_TRANSCRIPT.txt")
}
\arguments{
\item{dataDir}{directory where raw data files are located}

\item{outputDir}{directory where data files and transcript will be written}

\item{logLevel}{log level for output \code{['TRACE','DEBUG','INFO','WARNING','ERROR','FATAL']}}

\item{transcript}{filename for logging output}
}
\value{
Vector of file names created.
}
\description{
Smoke monitoring data from wrcc.com is downloaded, quality controlled
and converted in to a collection of .RData files, each containing a \code{ws_monitor}
for a particular UnitID.

By default, a transcript file will be generated containing information on the progress
of data downlaod, quality control and restructuring. More detailed information is
generated by setting \code{logLevel='DEBUG'}.

Setting \code{transcript=NULL} causes output to be sent to the console.
}
\examples{
\dontrun{
files <- wrcc_buildLibrary(dataDir='~/Data/WRCC', outputDir='~/Data/WRCC')
monitor_combineLibraryFiles('~/Data/WRCC', files, 'WRCC.RData')
wrcc <- <- wrcc_load(url='~/Data/WRCC/WRCC.RData')
monitor_leaflet(wrcc)
}
}
\references{
\href{http://usfs.wrcc.com}{Interagency Real Time Smoke Monitoring}
}
\seealso{
\link{monitor_combineLibraryFiles}
}
\keyword{WRCC}

