% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrcc_clustering.R
\name{wrcc_clustering}
\alias{wrcc_clustering}
\title{Add Clustering and Timezone information to a WRCC Dataframe}
\usage{
wrcc_clustering(df, clusterDiameter = 1000, minCount = 48,
  verbose = FALSE)
}
\arguments{
\item{df}{single site dataframe created by wrcc_qualityControl()}

\item{clusterDiameter}{diameter in meters used to determine the number of clusters (see description)}

\item{minCount}{minimum number of measurements for a location to be valid (see description)}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Clean dataframe of WRCC monitor data with clustering and timezone columns.
}
\description{
Clustering is used to assign individual measurements to deployment locations.

clusterRadius is compared with the output of cluster::pam(...)$clusinfo[,'diameter']
to determine the number of clusters.
}
\keyword{WRCC}

