% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMetaDataframe.R
\name{wrcc_createMetaDataframe}
\alias{wrcc_createMetaDataframe}
\title{Create Sites Metadata Dataframe}
\usage{
wrcc_createMetaDataframe(df, verbose = FALSE)
}
\arguments{
\item{df}{single site WRCC dataframe created by wrcc_clustering()}

\item{verbose}{logical flag to generate verbose output}
}
\value{
ws_monitor object with a unique `monitorID` for each unique location
}
\description{
Explain this function
}
\keyword{WRCC}

