% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createRawDataframe.R
\name{wrcc_createRawDataframe}
\alias{wrcc_createRawDataframe}
\title{Obtain WRCC Data and Parse into a Dataframe}
\usage{
wrcc_createRawDataframe(stationID = NULL, startdate = 20100101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  clusterDiameter = 1000,
  baseUrl = "http://www.wrcc.dri.edu/cgi-bin/wea_list2.pl", saveFile = NULL)
}
\arguments{
\item{stationID}{station identifier (will be upcased)}

\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{clusterDiameter}{diameter in meters used to determine the number of clusters (see \code{addClustering})}

\item{baseUrl}{base URL for data queries}

\item{saveFile}{optional filename where raw CSV will be written}
}
\value{
ws_monitor object with a unique `monitorID` for each unique deployment.
}
\description{
Obtains monitor data from a WRCC webservice and converts
it into a quality controlled, metadata enhanced "raw" dataframe
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download CSV text}
 \item{parse CSV text}
 \item{apply quality control}
 \item{apply clustering to determine unique deployments}
 \item{enhance metadata to include: elevation, timezone, state, country, site name}
}
}
\note{
The downloaded CSV may be saved to a local file by providing an argument to the \code{saveFile} parameter.
}
\seealso{
\code{\link{wrcc_downloadData}}

\code{\link{wrcc_parseData}}

\code{\link{wrcc_qualityControl}}

\code{\link{addClustering}}
}
\keyword{WRCC}

