% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{wrcc_downloadData}
\alias{wrcc_downloadData}
\title{Download Data from WRCC}
\usage{
wrcc_downloadData(stationID = NULL, startdate = 20100101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  url = "http://www.wrcc.dri.edu/cgi-bin/wea_list2.pl")
}
\arguments{
\item{stationID}{station identifier (will be upcased)}

\item{startdate}{integer or character representing start date as YYYYMMDD (GMT)}

\item{enddate}{integer or character representing end date as YYYYMMDD (GMT)}

\item{url}{base URL for data queries}
}
\value{
Dataframe of WRCC monitor data.
}
\description{
Create an HTTPPOST request to the WRCC data service to request
data from a particular station for the desired time period. Raw data are
returned as both a dataframe and a character string. Character data are in 
tab-separated-value format.
}
\examples{
\dontrun{
df <- wrcc_downloadData('SM16',startdate=20150701,enddate=20150930)
}
}
\references{
\href{http://www.wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
\keyword{WRCC}

