% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrcc_load.R
\name{wrcc_load}
\alias{wrcc_load}
\title{Load WRCC monitoring data}
\usage{
wrcc_load(startdate, enddate, monitorIDs = NULL, stateCodes = NULL,
  url = "http://smoke.airfire.org/RData/WRCC/WRCC_monitors.RData")
}
\arguments{
\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{monitorIDs}{optional vector of monitorIDs used to subset the data}

\item{stateCodes}{optional vector of stateCodes used to subset the data}

\item{url}{The location of the meta and data files (Default = 'http://smoke.airfire.org/RData/AirNowTech/')}
}
\value{
ws_monitor object with subsetted time, monitorIDs and parameter
}
\description{
When given the startdate, enddate, monitorIDs and parameter of interest, the function retrieves
data from the archive url and returns the subsetted ws_monitor object.
}
\examples{
\dontrun{
wrcc <- wrcc_load(20150901, 20150930)
wrcc_conus <- monitor_subset(wrcc, stateCodes=CONUS)
monitor_leaflet(wrcc_conus)
}
}
\keyword{wrcc}

