% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadLatest.R
\name{wrcc_loadLatest}
\alias{wrcc_loadLatest}
\title{Load Recent WRCC Monitoring Data}
\usage{

  wrcc_loadLatest(baseUrl = "https://haze.airfire.org/monitoring/WRCC/RData/")
}
\arguments{
\item{baseUrl}{location of the WRCC latest data file}
}
\value{
A \emph{ws_monitor} object with WRCC data.
}
\description{
Loads pre-generated .RData files containing the most recent WRCC data.

Avaialble RData and associated log files can be seen at:
\href{https://haze.airfire.org/monitoring/WRCC/RData/latest}{https://haze.airfire.org/monitoring/WRCC/RData/latest}
}
\examples{
\dontrun{
wrcc <- wrcc_loadLatest()
}
}
\seealso{
\code{\link{wrcc_load}}

\code{\link{wrcc_loadDaily}}
}
\keyword{WRCC}
