% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_parseData.R
\name{wrcc_parseData}
\alias{wrcc_parseData}
\title{Parse WRCC Data String}
\usage{
wrcc_parseData(fileString)
}
\arguments{
\item{fileString}{character string containing WRCC data}
}
\value{
Dataframe of WRCC raw monitor data.
}
\description{
Raw character data from WRCC are parsed into a dataframe.
The incoming \code{fileString}
can be read in directly from WRCC using \code{wrcc_downloadData()} or from a local
file using \code{readr::read_file()}.

The type of monitor represented by this fileString is inferred from the column names
using \code{wrcc_identifyMonitorType()} and appropriate column types are assigned.
The character data are then read into a dataframe and augmented in the following ways:
\enumerate{
\item{Longitude, Latitude and any System Voltage values, which are only present in GPS timestamp rows, are
propagated foward using a last-observation-carry-forward algorithm'}
\item{Longitude, Latitude and any System Voltage values, which are only present in GPS timestamp rows, are
propagated backwords using a first-observation-carry-backward algorithm'}
\item{GPS timestamp rows are removed'}
}
}
\examples{
\dontrun{
fileString <- wrcc_downloadData('SM16',startdate=20150701,enddate=20150930)
df <- wrcc_parseData(fileString)
}
}
\references{
\href{http://www.wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
\keyword{WRCC}

