% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrcc_qualityControl.R
\name{wrcc_qualityControl}
\alias{wrcc_qualityControl}
\title{Apply Quality Control to Raw WRCC Dataframe}
\usage{
wrcc_qualityControl(df, verbose = FALSE)
}
\arguments{
\item{df}{single site dataframe created by wrcc_readData()}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Cleaned up dataframe of WRCC monitor data.
}
\description{
Perform various QC measures on WRCC data.
}
\keyword{WRCC}

