% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrcc_readData.R
\name{wrcc_readData}
\alias{wrcc_readData}
\title{Read and Parse Single Monitor Data from WRCC}
\usage{
wrcc_readData(file, verbose = FALSE)
}
\arguments{
\item{file}{either a path to a WRCC data file the contents of the file as a character string}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Dataframe of WRCC monitor data.
}
\description{
A WRCC monitor data file is read in and parsed.
}
\keyword{WRCC}

