% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrccDump_parseData.R
\name{wrccDump_parseData}
\alias{wrccDump_parseData}
\title{Parse WRCC Dump File String}
\usage{
wrccDump_parseData(fileString)
}
\arguments{
\item{fileString}{character string containing WRCC dump file}
}
\value{
List of dataframes of WRCC raw monitor data from multiple monitors.
}
\description{
Raw character data from WRCC are parsed into a dataframe.
The incoming \code{fileString} can be read from a local \code{SmokeMon.csv}
file using \code{readr::read_file()}.

This function is intended for internal use at the Pacific Wildland
Fire Sciences Lab.
}
\keyword{WRCC}

