% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAQIStackedBar.R
\name{addAQIStackedBar}
\alias{addAQIStackedBar}
\title{Create Stacked AQI Bar}
\usage{
addAQIStackedBar(width = 0.01, height = 1, pos = "left")
}
\arguments{
\item{width}{width of the bar as a fraction of the width of the plot area (default = .02)}

\item{height}{height of the bar as a fraction of the height of the plot area (default = 1)}

\item{pos}{position of the stacked bar. Either 'left' or 'right'}
}
\value{
Stacked AQI Bar
}
\description{
Draws a stacked bar indicating AQI levels on one side of a plot
}
\keyword{plotting}
