% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBullseye.R
\name{addBullseye}
\alias{addBullseye}
\title{Add a Bullseye at the Specified Location}
\usage{
addBullseye(x, y)
}
\arguments{
\item{x}{x-location on plot}

\item{y}{y-locaiton on plot}
}
\description{
Draws a bullseye with concentric rings of black and white.
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
KingFire <- monitor_subsetByDistance(airnow, lon=-120.604, lat=38.782, radius=50)
map('county',region='ca', col='gray80')
monitor_map(KingFire, add=TRUE)
addBullseye(-120.604, 38.782)
text(-120.604, 38.782, "King Fire", pos=4, font=2)
}
}
\seealso{
\link{addShadedNights}
}
\keyword{plotting}

