% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBullseye.R
\name{addBullseye}
\alias{addBullseye}
\title{Add a Bullseye at the Specified Location}
\usage{
addBullseye(x, y)
}
\arguments{
\item{x}{x-location on plot}

\item{y}{y-locaiton on plot}
}
\description{
Draws a bullseye with concentric rings of black and white.
}
\examples{
\dontrun{
setSmokeDataDir("~/Data/Smoke/")
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2015070600)
xlim <- c(-124.801, -122.801)
ylim <- c(47.004, 48.404)
grid_map(bs, xlim = xlim, ylim = ylim)
addBullseye(-123.801, 47.704)
}
}
\seealso{
\link{addShadedNights}
}
\keyword{plotting}

