% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addClustering.R
\name{addClustering}
\alias{addClustering}
\title{Add Clustering Information to a Dataframe}
\usage{
addClustering(df, clusterDiameter = 1000, minCount = 48,
  lonVar = "longitude", latVar = "latitude", maxClusters = 50,
  verbose = FALSE)
}
\arguments{
\item{df}{dataframe with geolocation information (e.g. created by wrcc_qualityControl() or airsis_qualityControl)}

\item{clusterDiameter}{diameter in meters used to determine the number of clusters (see description)}

\item{minCount}{minimum number of measurements for a location to be valid (see description)}

\item{lonVar}{name of longitude variable in the incoming dataframe}

\item{latVar}{name of the latitude variable in the incoming dataframe}

\item{maxClusters}{maximum number of clusters to try}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Input dataframe with additional columns: deploymentID, medoidLon, mediodLat.
}
\description{
Clustering is used to assign individual measurements to deployment locations.

clusterRadius is compared with the output of cluster::pam(...)$clusinfo[,'av_diss']
to determine the number of clusters.
}
\keyword{internal}

