% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLegend.R
\name{addLegend}
\alias{addLegend}
\title{Add a Legend to a Map of Monitoring Stations}
\usage{
addLegend(x = "topright", y = NULL, col = rev(AQI$colors),
  legend = rev(AQI$names), pch = 16, title = "Air Quality Index", ...)
}
\arguments{
\item{x}{x coordinate passed on to the legend() command}

\item{y}{y coordinate passed on to the legend() command}

\item{col}{the color for points/lines in the legend}

\item{legend}{a character vector to be shown in the legend}

\item{pch}{plotting symbols in the legend}

\item{title}{title for the legend}

\item{...}{additional arguments to be passed to legend()}
}
\description{
This function is a convenience wrapper around graphics::legend(). It will show the AQI colors and 
names by default if \code{col} and \code{legend} are not specified.
}
\examples{
\dontrun{
airnow <- airnow_load(20150901, 20150930)
nw <- monitor_subset(airnow, stateCodes=c('WA','OR'), countryCodes="US")
monitorMap(nw, showLegend=FALSE)
addLegend(x="left", bty="n", cex=0.8, inset=c(-0.2,0), xpd=TRUE)
}
}
\keyword{legend}

