% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addShadedNights.R
\name{addShadedNights}
\alias{addShadedNights}
\title{Add Nighttime Shading on an Existing Plot}
\usage{
addShadedNights(timeInfo, col = adjustcolor("black", 0.1))
}
\arguments{
\item{timeInfo}{dataframe with local time, sunrise, and sunset}

\item{col}{color used to shade nights -- defaults to \code{adjustcolor('black',0.2)}}
}
\description{
Draw shading rectangles on a plot to indicate nighttime hours.
}
\examples{
\dontrun{
setSmokeDataDir("~/Data/Smoke")
bs <- bluesky_load(model='PNW-1.33km', modelRun=2015070600)
targetLon <- -123.801 ; targetLat <- 47.704
timeInfo <- timeInfo(bs$time, targetLon, targetLat)
bsLocal <- grid_subsetByDistance(bs, targetLon, targetLat, 10)
monitor_timeseriesPlot(bsLocal, col='red')
addShadedNights(timeInfo)
}
}

