% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airnow_createDataDataframe.R
\name{airnow_createDataDataframe}
\alias{airnow_createDataDataframe}
\title{Return Reshaped, Monthly Dataframes of AirNow Data}
\usage{
airnow_createDataDataframe(user, pass, yearMonth, tries = 6,
  verbose = FALSE, debug = FALSE)
}
\arguments{
\item{user}{user name}

\item{pass}{pass}

\item{yearMonth}{desired year and month  (integer or character representing YYYYMM)}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical requesting verbose output from libcurl}

\item{debug}{logical specyfing whether to print out additional debugging lines}
}
\value{
Both raw and processed list of data frames where each of them is the data table for the unique parameter (e.g: PM2.5, NOX)
}
\description{
The airnow_createDataDataframe() function uses the \link{airnow_downloadData} function
to download monthly dataframes of AirNow data and restructures that data into a format that is compatible
with the PWFSLSmoke package \emph{ws_monitor} data model.

The function also has built in
post-processing script that returns both raw and processed data.
}
\examples{
\dontrun{
AirnowData <- airnow_createDataDataframe(user, pass, 201507)
}
}
\seealso{
\link{airnow_downloadData}

\link{airnow_downloadMonthlyData}
}
\keyword{AirNow}

