% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_downloadData.R
\name{airnow_downloadData}
\alias{airnow_downloadData}
\title{Download and Aggregate Multiple Hourly Data Files from AirNow}
\usage{
airnow_downloadData(user = "", pass = "", parameters = NULL,
  startdate = "", hours = 24, tries = 6, verbose = FALSE, curl = NULL)
}
\arguments{
\item{user}{user name}

\item{pass}{password}

\item{parameters}{vector of names of desired pollutants or NULL for all pollutants}

\item{startdate}{desired staring date (integer or character representing YYYYMMDD[HH])}

\item{hours}{desired number of hours of data to assemble}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical flag to generate verbose web connection output}

\item{curl}{previously initialized CURL context/handle (see RCurl::getURL())}
}
\value{
Dataframe of aggregated AirNow data.
}
\description{
This function makes repeated calls to \link{airnow_downloadHourlyData}
to obtain data from AirNow. A single curl handle is created and reused to improve 
peformance and avoid errors like "Maximum login connections". All data obtained are then
combined into a single dataframe and returned.

Parameters included in AirNow data include at least the following list:
\enumerate{
\item{BARPR}
\item{BC}
\item{CO}
\item{NO}
\item{NO2}
\item{NO2Y}
\item{NO2X}
\item{NOX}
\item{NOOY}
\item{OC}
\item{OZONE}
\item{PM10}
\item{PM2.5}
\item{PRECIP}
\item{RHUM}
\item{SO2}
\item{SRAD}
\item{TEMP}
\item{UV-AETH}
\item{WD}
\item{WS}
}

Passing a vector of one ore more of the above names as the \code{parameters} argument will cause the resulting 
dataframe to be filtered to contain only records for those parameters.
}
\examples{
\dontrun{
df <- airnow_downloadData(USER, PASS, "PM2.5", 2015070112, hours=48)
}
}
\seealso{
\link{airnow_createDataDataframes}

\link{airnow_downloadHourlyData}
}
\keyword{AirNow}

