% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airnow_downloadHourlyData.R
\name{airnow_downloadHourlyData}
\alias{airnow_downloadHourlyData}
\title{Download Hourly Data from AirNow}
\usage{
airnow_downloadHourlyData(user = "", pass = "", datestamp = "",
  tries = 6, verbose = FALSE,
  baseUrl = "ftp.airnowapi.org/HourlyData/Archive/", curl = NULL)
}
\arguments{
\item{user}{username}

\item{pass}{password}

\item{datestamp}{integer or character representing YYYYMMDDHH}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical requesting verbose output from libcurl}

\item{baseUrl}{base URL for archived hourly data}

\item{curl}{previously initialized CURL context/handle (see RCurl::getURL())}
}
\value{
Data frame of the AirNow hourly data.
}
\description{
With the given username and password, airnow_downloadHourlyData() uses the \code{datesteamp}
argument to create a URL into the Airnow archive at \url{'ftp.airnowapi.org/HourlyData/Archive}.
The associated ASCII hourly data file is downloaded and converted into a dataframe.
}
\examples{
\dontrun{
df <- airnow_downloadHourlyData(USER, PASS, 2015070112)
}
}
\keyword{AirNow}
\keyword{internal}

