% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_downloadMonthlyData.R
\name{airnow_downloadMonthlyData}
\alias{airnow_downloadMonthlyData}
\title{Download and Aggregate a Month of Hourly Data Files from AirNow}
\usage{
airnow_downloadMonthlyData(user = "", pass = "", yearMonth = "",
  tries = 6, verbose = FALSE)
}
\arguments{
\item{user}{user name}

\item{pass}{password}

\item{yearMonth}{desired year and month  (integer or character representing YYYYMM)}

\item{tries}{number of download attempts in the face of timeouts}

\item{verbose}{logical flag to generate verbose output}
}
\value{
Dataframe of aggregated AirNow data.
}
\description{
This function is a convenience wrapper for \link{airnow_downloadData}
which automatically determines the number of hours of data to download for a particular month.
}
\examples{
\dontrun{
df <- airnow_downloadMonthlyData(USER, PASS, 201507)
}
}
\seealso{
\link{airnow_downloadData}
}
\keyword{AirNow}

