% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_load.R
\name{airnow_load}
\alias{airnow_load}
\title{Load Processed AirNow Monitoring Data}
\usage{
airnow_load(
  year = 2017,
  month = NULL,
  parameter = "PM2.5",
  baseUrl = "https://haze.airfire.org/monitoring/AirNow/RData/"
)
}
\arguments{
\item{year}{desired year (integer or character representing YYYY)}

\item{month}{desired month (integer or character representing MM)}

\item{parameter}{parameter of interest}

\item{baseUrl}{base URL for AirNow meta and data files}
}
\value{
A \emph{ws_monitor} object with AirNow data.
}
\description{
Please use \code{\link{airnow_loadAnnual}} instead of this
function. It will soon be deprecated.
}
\keyword{AirNow}
