% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_qualityControl.R
\name{airnow_qualityControl}
\alias{airnow_qualityControl}
\title{Apply Quality Control to AirNow Dataframe}
\usage{
airnow_qualityControl(df, limits = c(-Inf, Inf))
}
\arguments{
\item{df}{multi-site restructured dataframe created within \code{airnow_createDataDataframe()}}

\item{limits}{lo and hi range of valid values}
}
\value{
Cleaned up dataframe of AIRSIS monitor data.
}
\description{
Perform range validation on AirNow data. This function also replaces values 
of \code{-999} with \code{NA}.
}
\seealso{
\link{airnow_createDataDataframes}
}
\keyword{AirNow}
