% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_createDataDataframe.R
\name{airsis_createDataDataframe}
\alias{airsis_createDataDataframe}
\title{Create AIRSIS Data Dataframe}
\usage{
airsis_createDataDataframe(df, meta)
}
\arguments{
\item{df}{single site AIRSIS dataframe created by airsis_clustering()}

\item{meta}{AIRSIS meta dataframe created by airsis_createMetaDataframe()}
}
\value{
A 'data' dataframe for use in a ws_monitor object.
}
\description{
After quality control has been applied to an AIRSIS dataframe,
we can extract the PM2.5 values and store them in a 'data' dataframe
organized as hour-by-deployment (aka hour-by-site).

The first column of the returned dataframe is named \code{'datetime'} and
contains a \code{POSIXct} time in UTC. Additional columns contain data
for each separate deployment of a monitor.
}
\keyword{AIRSIS}

