% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_identifyMonitorType.R
\name{airsis_identifyMonitorType}
\alias{airsis_identifyMonitorType}
\title{Identify AIRSIS Monitor Type}
\usage{
airsis_identifyMonitorType(df)
}
\arguments{
\item{df}{dataframe or raw character string containing AIRSIS data}
}
\value{
List including \code{monitorType}, \code{rawNames}, \code{columnNames} and \code{columnTypes}.
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

