% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_load.R
\name{airsis_load}
\alias{airsis_load}
\title{Load Processed AIRSIS Monitoring Data}
\usage{
airsis_load(
  year = 2017,
  baseUrl = "https://haze.airfire.org/monitoring/AIRSIS/RData/"
)
}
\arguments{
\item{year}{desired year (integer or character representing YYYY)}

\item{baseUrl}{base URL for AIRSIS meta and data files}
}
\value{
A \emph{ws_monitor} object with AIRSIS data.
}
\description{
Please use \code{\link{airsis_loadAnnual}} instead of this
function. It will soon be deprecated.
}
\keyword{AIRSIS}
