% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_load.R
\name{airsis_load}
\alias{airsis_load}
\title{Load AIRSIS Monitoring Data}
\usage{
airsis_load(startdate, enddate, monitorIDs = NULL, stateCodes = NULL,
  url = "http://smoke.airfire.org/RData/AIRSIS/AIRSIS_monitors.RData")
}
\arguments{
\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{monitorIDs}{optional vector of monitorIDs used to subset the data}

\item{stateCodes}{optional vector of stateCodes used to subset the data}

\item{url}{The location of the meta and data files (Default = 'http://smoke.airfire.org/RData/AirNowTech/')}
}
\value{
ws_monitor object with subsetted time, monitorIDs and parameter
}
\description{
When given the startdate, enddate, monitorIDs and parameter of interest, the function retrieves
data from the archive url and returns the subsetted ws_monitor object.
}
\examples{
\dontrun{
airsis <- airsis_load(2010901, 20141130)
airsis_conus <- monitor_subset(airsis, stateCodes=CONUS)
monitor_leaflet(airsis_conus)
}
}
\keyword{airsis}

