% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_load.R
\name{airsis_load}
\alias{airsis_load}
\title{Load AIRSIS Monitoring Data}
\usage{
airsis_load(startdate = 20000101, enddate = strftime(lubridate::now(),
  "\%Y\%m\%d", tz = "GMT"), monitorIDs = NULL, stateCodes = NULL,
  url = "https://smoke.airfire.org/RData/AIRSIS/AIRSIS_monitors.RData")
}
\arguments{
\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{monitorIDs}{optional vector of monitorIDs used to subset the data}

\item{stateCodes}{optional vector of stateCodes used to subset the data}

\item{url}{location of the meta and data files (url or local file)}
}
\value{
A ws_monitor object with AIRSIS data.
}
\description{
When given the startdate, enddate, monitorIDs and parameter of interest, the function retrieves 
data from the archive url and returns the subsetted ws_monitor object.
}
\examples{
\dontrun{
airsis <- airsis_load(20140901, 20141130)
}
}
\keyword{AIRSIS}

