% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsis_loadLatest.R
\name{airsis_loadLatest}
\alias{airsis_loadLatest}
\title{Load Recent AIRSIS Monitoring Data}
\usage{
airsis_loadLatest(baseUrl = "http://tools.airfire.org/monitoring/v3/data")
}
\arguments{
\item{baseUrl}{The location of the meta and data files}
}
\value{
ws_monitor object with subsetted time, monitorIDs and parameter
}
\description{
The most recent 10 days of AIRSIS data are updated in quasi-real time
at PWFSL and can be loaded with this function.
}
\examples{
\dontrun{
airsis <- airsis_loadLatest()
}
}
\keyword{airsis}

