% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airsisDump_parseData.R
\name{airsisDump_parseData}
\alias{airsisDump_parseData}
\title{Parse AIRSIS Data String}
\usage{
airsisDump_parseData(fileString)
}
\arguments{
\item{fileString}{character string containing AIRSIS data as a csv}
}
\value{
Dataframe of AIRSIS raw monitor data.
}
\description{
Raw character data from AIRSIS are parsed into a dataframe.
The incoming \code{fileString}
can be read in directly from AIRSIS using \code{airsis_downloadData()} or from a local
file using \code{readr::read_file()}.

The type of monitor represented by this fileString is inferred from the column names
using \code{airsis_identifyMonitorType()} and appropriate column types are assigned.
The character data are then read into a dataframe and augmented in the following ways:
\enumerate{
\item{Longitude, Latitude and any System Voltage values, which are only present in GPS timestamp rows, are
propagated foward using a last-observation-carry-forward algorithm'}
\item{Longitude, Latitude and any System Voltage values, which are only present in GPS timestamp rows, are
propagated backwords using a first-observation-carry-backward algorithm'}
\item{GPS timestamp rows are removed'}
}
}
\examples{
\dontrun{
fileString <- airsis_downloadData(USFS',unitID='1026',startdate=20150701,enddate=20151231)
df <- airsis_parseData(fileString)
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
\keyword{AIRSIS}

