% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{AQI}
\alias{AQI}
\title{Official Air Quality Index Levels, Names and Colors}
\format{A list with five elements}
\usage{
AQI
}
\description{
Official AQI levels, names and colors are provided in a list for easy coloring and labeling.
}
\details{
AQI breaks and associated names and colors



AQI breaks were obtained from \url{http://www.arb.ca.gov/carpa/toolkit/data-to-mes/wildfire-smoke-guide.pdf}

AQI colors were copied from \url{http://aqicn.org/faq/2013-09-09/revised-pm25-aqi-breakpoints/}
}
\note{
The low end of each break category is used as the breakpoint.
}
\keyword{datasets}

