% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{Carmel_Valley}
\alias{Carmel_Valley}
\title{Carmel Valley Example Dataset}
\format{A list with two elements}
\description{
In August of 2016, the Soberanes fire in California burned along the Big Sur
coast. It was at the time the most expensive wildifre in US history. This dataset contains
PM2.5 monitoring data for the monitor in Carmel Valley which shows heavy smoke
as well as strong diurnal cycles associated with sea breezes. Data are stored
as a \emph{ws_monitor} object and are used in some examples in the package
documentation.
}
\details{
Carmel Valley example dataset
}
\keyword{datasets}
