% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{CONUS}
\alias{CONUS}
\title{CONUS State Codes}
\format{A vector with 48 elements}
\usage{
CONUS
}
\description{
State codes for the 48 states +DC that make up the Continental US
}
\details{
CONUS state codes
}
\keyword{datasets}
