% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculate the Distance Between Points}
\usage{
distance(targetLon, targetLat, lons, lats)
}
\arguments{
\item{targetLon}{longitude (decimal degrees) of the point from which distances are calculated}

\item{targetLat}{latitude (decimal degrees) of the point from which distances are calculated}

\item{lons}{vector of longitudes (decimal degrees) for which a distance is calculated}

\item{lats}{vector of latitudes (decimal degrees) for which a distance is calculated}
}
\value{
Vector of distances in km.
}
\description{
This funciton uses the Haversine forumula for calculating great
circle distances between points. This formula is purpoted to work better
than the spherical law of cosines for very short distances.
}
\references{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

