% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epa_createMonitorObject.R
\name{epa_createMonitorObject}
\alias{epa_createMonitorObject}
\title{Download and Convert Hourly EPA Air Quality Data}
\usage{
epa_createMonitorObject(parameterName = "PM2.5", parameterCode = 88101,
  year = NULL, verbose = TRUE,
  baseUrl = "http://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/")
}
\arguments{
\item{parameterName}{name of parameterName.}

\item{parameterCode}{specific parameter code (e.g. PM2.5 could be 88101 or 88502)}

\item{year}{year}

\item{verbose}{logical flag to generate verbose output}

\item{baseUrl}{base URL for archived hourly data}
}
\value{
Character string name of the generated .RData file.
}
\description{
Convert EPA data into a ws_monitor object, ready for use with all monitor_~ functions.
}
\note{
Before running this function you must first enable spatial data capabilities as in the example.
}
\examples{
\dontrun{
library(PWFSLSmoke)
setSmokeDataDir('~/Data/Smoke')
library(MazamaSpatialUtils)
setSpatialDataDir('~/Data/Spatial')
loadSpatialData('NaturalEarthAdm1') # stateCodes dataset
file <- epa_createMonitorObject("PM2.5", 88101, 2015, verbose=TRUE)
mon <- get(load(file))
}
}
\references{
\href{http://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/download_files.html}{EPA data files}

\href{http://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/FileFormats.html#_format_3}{file format description}
}
\keyword{ws_monitor}

