% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epa_downloadData.R
\name{epa_downloadData}
\alias{epa_downloadData}
\title{Download hourly air-qiality data from a given year}
\usage{
epa_downloadData(parameterName = NULL, parameterCode = NULL, year = NULL,
  baseUrl = "https://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/")
}
\arguments{
\item{parameterName}{pollutant name}

\item{parameterCode}{a character string of pollutant code}

\item{year}{a numeric year value}

\item{baseUrl}{base URL for archived daily data}
}
\value{
Dataframe of EPA data.
}
\description{
This function downloads data from the EPA

Available parameters include:
\enumerate{
\item{Ozone}
\item{SO2}
\item{CO}
\item{NO2}
\item{PM2.5}
\item{PM10}
\item{Wind}
\item{Temperatue}
\item{Barometric_Pressure}
\item{RH_and_Dewpoint}
\item{HAPs}
\item{VOCs}
\item{NONOxNOy}
}

Available parameter codes include:
\enumerate{
\item{44201}{ --Ozone}
\item{42401}{ --SO2}
\item{42101}{ --CO}
\item{42602}{ --NO2}
\item{88101}{ --PM2.5}
\item{88502}{ --PM2.5}
\item{81102}{ --PM10}
\item{SPEC}{ --PM2.5}
\item{WIND}{ --Wind}
\item{TEMP}{ --Temperature}
\item{PRESS}{ --Barometric Pressure}
\item{RH_DP}{ --RH and dewpoint}
\item{HAPS}{ --HAPs}
\item{VOCS}{ --VOCs}
\item{NONOxNOy}
}
}
\examples{
\dontrun{
df <- epa_downloadData(parameterName="PM2.5", parameterCode="88101", year=2016)
}
}
\keyword{EPA}

