% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_load.R
\name{epa_load}
\alias{epa_load}
\title{Load Processed EPA Monitoring Data}
\usage{
epa_load(
  year = strftime(lubridate::now(tzone = "UTC"), "\%Y", tz = "UTC"),
  parameterCode = "88101",
  baseUrl = "https://haze.airfire.org/monitoring/EPA/RData/"
)
}
\arguments{
\item{year}{desired year (integer or character representing YYYY)}

\item{parameterCode}{pollutant code}

\item{baseUrl}{base URL for EPA .RData files}
}
\value{
A \emph{ws_monitor} object with EPA data for an entire year.
}
\description{
Please use \code{\link{airsis_loadAnnual}} instead of this
function. It will soon be deprecated.
}
\keyword{EPA}
