% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriMap_getMap.R
\name{esriMap_getMap}
\alias{esriMap_getMap}
\title{Create a RGB spatial raster from arcGIS REST}
\usage{
esriMap_getMap(centerLon, centerLat, mapType = "worldStreetMap", zoom = 12,
  width = 640, height = 640, crs = sp::CRS("+init=epsg:3857"))
}
\arguments{
\item{centerLon}{map center longitude}

\item{centerLat}{map center latitude}

\item{mapType}{map type. natGeo, worldStreetMap, worldTopoMap, satellite, or deLorme. Also accepts
map server identity, found at \url{http://resources.arcgis.com/en/help/arcgis-rest-api/index.html#/Basemaps/02r3000001mt000000/}}

\item{zoom}{map zoom level; corresponds to googleMaps zoom level}

\item{width}{width of image, in pixels}

\item{height}{height of image, in pixels}

\item{crs}{object of class CRS. The Coordinate Reference System (CRS) for the returned map. If the CRS of the downloaded
map does not match, it will be projected to the specified CRS using \code{raster::projectRaster}.}
}
\value{
A rasterBrick object which can be plotted with \code{raster::plotRGB} and serve as a base plot.
}
\description{
Creates a rasterBrick using the \pkg{raster} package with layers for red, green, and blue color intensity.
}
\examples{
map <- esriMap_getMap(-122.3318, 47.668)
raster::plotRGB(map)
map <- esriMap_getMap(-122.3318, 47.668, crs = sp::CRS("+init=epsg:4326"))
raster::plotRGB(map)
}
\references{
\url{http://resources.arcgis.com/en/help/arcgis-rest-api/index.html#/Export_Map/02r3000000v7000000/}
}
\keyword{plotting}
