% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\name{getSmokeDataDir}
\alias{getSmokeDataDir}
\title{Get Package Data Directory}
\usage{
getSmokeDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the package data directory where smoke data is located.
}
\seealso{
\link{SmokeDataDir}

\link{setSmokeDataDir}
}
\keyword{environment}

