% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFile.R
\name{loadDataFile}
\alias{loadDataFile}
\title{Load data from URL or local file}
\usage{
loadDataFile(
  filename = NULL,
  baseUrl = "https://haze.airfire.org/monitoring/latest/RData",
  dataDir = NULL
)
}
\arguments{
\item{filename}{Name of the data file to be loaded.}

\item{baseUrl}{Base URL for data files.}

\item{dataDir}{Local directory containing data files.}
}
\value{
A \emph{ws_monitor} object.
}
\description{
Loads pre-generated .RData files
}
\keyword{internal}
