% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pylogger.R
\docType{data}
\name{logLevels}
\alias{DEBUG}
\alias{ERROR}
\alias{FATAL}
\alias{INFO}
\alias{TRACE}
\alias{WARN}
\alias{logLevels}
\title{Log Levels}
\format{\preformatted{ Named int 1
 - attr(*, "names")= chr "FATAL"
}}
\usage{
FATAL
}
\description{
Log levels matching those found in \pkg{futile.logger}.
}
\keyword{datasets}

