% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylogger.R
\docType{data}
\name{logLevels}
\alias{logLevels}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{Log Levels}
\format{An object of class \code{integer} of length 1.}
\usage{
FATAL
}
\description{
Log levels matching those found in \pkg{futile.logger}.
Available levels include:

\code{FATAL ERROR WARN INFO DEBUG TRACE}
}
\keyword{datasets}
