% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_aqi.R
\name{monitor_aqi}
\alias{monitor_aqi}
\title{Calculate AQI for ws_monitor Object}
\usage{
monitor_aqi(ws_monitor, parameter = "pm25", hour = 24)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{parameter}{pollutant type}

\item{hour}{the number of hours for concentration values to be averaged}
}
\value{
A ws_monitor object with data replaced by AQI values.
}
\description{
AQI values are calculated for data in the ws_monitor object for different pollutants.

Available combinations of \code{parameter} and \code{hour} are:
\enumerate{
\item{parameter="o3", hour=8}
\item{parameter="o3", hour=1}
\item{parameter="pm25, hour=24}
\item{parameter="pm10, hour=24}
\item{parameter="co", hour=8}
\item{parameter="so2", hour=1}
\item{parameter="no2", hour=1}
}

See references for calculation details.
}
\examples{
\dontrun{
airnow <- airnow_load(20161001, 20161031)
airnow_AQI <- monitor_aqi(airnow)
}

}
\references{
\url{https://www3.epa.gov/airnow/aqi-technical-assistance-document-may2016.pdf}
}
\keyword{AQI}

