% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_collapse.R
\name{monitor_collapse}
\alias{monitor_collapse}
\title{Collapse a ws_monitor Object into a ws_monitor Object with a Single Monitor}
\usage{
monitor_collapse(ws_monitor, lon = NULL, lat = NULL,
  monitorID = "generated_id", FUN = mean, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{lon}{The custom longitude of the collapsed monitoring station. (Default = mean of the longitudes)}

\item{lat}{The custom latitude of the collapsed monitoring station. (Default = mean of the latitudes)}

\item{monitorID}{The custom monitor ID of the collapsed monitoring station. (Default = 'generated_id')}

\item{FUN}{The function to be applied to all the monitors at a single time index. (Default = mean)}

\item{...}{Any additional argument to be passed on to the \code{apply()} function}
}
\value{
A ws_monitor object with meta and data that corresponds to the collapsed single monitor
}
\description{
Collapses all the monitors from a given ws_monitor object into a single-monitor ws_monitor object using
the function provided in the \code{FUN} argument. The collapse applies to both ws_monitor$meta and ws_monitor$data.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke/')
# Smokey days in the Okanagan
airnow <- airnow_load(startdate=2015080100, enddate=2015083100)
Omak <- monitor_subset(airnow, monitorIDs='530470013')
bs <- bluesky_load(modelRun=2015081900)
bs_near_Omak <- grid_subsetByDistance(bs, lon=Omak$meta$longitude, lat=Omak$meta$latitude, radius=5)
bs_Omak <- monitor_collapse(bs_near_Omak, monitorID='bs_Omak')
}
}
\keyword{ws_monitor}

