% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_collapse.R
\name{monitor_collapse}
\alias{monitor_collapse}
\title{Collapse a ws_monitor Object into a ws_monitor Object with a Single Monitor}
\usage{
monitor_collapse(ws_monitor, lon = NULL, lat = NULL,
  monitorID = "generated_id", FUN = mean, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{lon}{longitude of the collapsed monitoring station. (Default = mean of the longitudes)}

\item{lat}{latitude of the collapsed monitoring station. (Default = mean of the latitudes)}

\item{monitorID}{monitor ID of the collapsed monitoring station. (Default = 'generated_id')}

\item{FUN}{function to be applied to all the monitors at a single time index. (Default = mean)}

\item{...}{additional arguments to be passed on to the apply() function}
}
\value{
A ws_monitor object with meta and data that corresponds to the collapsed single monitor
}
\description{
Collapses all the monitors from a given ws_monitor object into a single-monitor ws_monitor object using
the function provided in the \code{FUN} argument. The collapse applies to both ws_monitor$meta and ws_monitor$data.

Intended for use with gridded model data, rather than actual monitor data itself (meta data is largely chopped; NAs propagate).
}
\keyword{internal}

