% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_combine.R
\name{monitor_combine}
\alias{monitor_combine}
\title{Combine ws_monitor Objects}
\usage{
monitor_combine(ws_monitor1, ws_monitor2)
}
\arguments{
\item{ws_monitor1}{data list of class \code{ws_monitor}}

\item{ws_monitor2}{a \code{ws_monitor} object to be merged with ws_monitor1}
}
\value{
A ws_monitor object where the ws_monitor1 and ws_monitor2 are combined
}
\description{
When given two ws_monitor objects, the function generates a single ws_monitor object
by merging both of their meta and data
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke/')
# Smokey days in the Okanagan
airnow <- airnow_load(startdate=2015081900, enddate=2015082200)
Omak <- monitor_subset(airnow, monitorIDs='530470013')
bs <- bluesky_load(modelRun=2015081900)
bs_near_Omak <- grid_subsetByDistance(bs, lon=Omak$meta$longitude, lat=Omak$meta$latitude, radius=5)
bs_Omak <- monitor_collapse(bs_near_Omak, monitorID='bs_Omak')
Omak_both <- monitor_combine(Omak, bs_Omak)
monitor_dygraph(Omak_both)
}
}
\keyword{ws_monitor}

