% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_currentData.R
\name{monitor_currentData}
\alias{monitor_currentData}
\title{Create a dataframe of current monitor data}
\usage{
monitor_currentData(ws_monitor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
A tibble of current status data with the following columns: 
monitorID, pm25, nowcast, yesterdayAqi, lastValidUpdateTime, processingTime.
}
\description{
Extracts current status data from a ws_monitor object. This data includes the following:
\itemize{
\item{monitorID - the PWFSL monitor ID}
\item{lastValidUpdateTime - the UTC timestamp corresponding to the last valid pm25 datum}
\item{processingTime - the UTC timestamp that the function is run}
\item{latency - the difference between processingTime and lastValidUpdateTime, floored to the hour}
\item{PM2.5_nowcast - the NowCast value at lastValidUpdateTime}
\item{PM2.5_latest_1 - the PM2.5 value at lastValidUpdateTime (should never be null)}
\item{PM2.5_latest_3 - the mean of the three hours preceeding lastValidUpdateTime}
\item{PM2.5_yesterday - the 24-hour mean from midnight-to-midnight of the day before processingTime}
}
}
\examples{
wa <- monitor_subset(Northwest_Megafires)
wa_current <- monitor_currentData(wa)
}
\keyword{ws_monitor}
