% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_currentData.R
\name{monitor_currentData}
\alias{monitor_currentData}
\title{Create a dataframe of current monitor data}
\usage{
monitor_currentData(ws_monitor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
A tibble of current status data with the following columns: 
monitorID, pm25, nowcast, yesterdayAqi, lastValidUpdateTime, processingTime.
}
\description{
Extracts current status data from a ws_monitor object. This data includes the following:
\itemize{
\item{monitorID - the PWFSL monitor ID}
\item{lastValidUpdateTime - UTC POSIXct time corresponding to the last valid pm25 datum}
\item{processingTime - UTC POSIXct when the function is run (= 'now')}
\item{latency - (difference between processingTime -1 hr) and lastValidUpdateTime, floored to the hour}
\item{PM2.5_nowcast - NowCast value at lastValidUpdateTime}
\item{PM2.5_latest_1 - PM2.5 value at lastValidUpdateTime (should never be null)}
\item{PM2.5_latest_3 - mean of the three hours preceeding lastValidUpdateTime}
\item{PM2.5_yesterday - local time midnight-to-midnight 24-hour mean for the day prior to processingTime}
}
}
\note{
Data are assigned to the beginning of the hour they represent. So a PM2.5 value assigned to 2pm
will represent data averaged over the period 14:00 - 14:59. This is in keeping with a day representing
00:00 - 23:59 and a month representing the beginning of the month until the end of the month.

Because of this, data for 2pm should never be available until just after 3pm. If it is currently 3:15pm
then we need to subtract 1 hour from the current \code{processingTime} before subtracting the \code{lastValidUpdateTime}
to generate the \code{latency}.
}
\examples{
\dontrun{
wa <- loadLatest() \%>\% monitor_subset(stateCodes = 'WA')
wa_current <- monitor_currentData(wa)
}
}
\keyword{ws_monitor}
