% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_currentData.R
\name{monitor_currentData}
\alias{monitor_currentData}
\title{Create a dataframe of current monitor data}
\usage{
monitor_currentData(ws_monitor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}
}
\value{
A tibble of "latest" data and associated timing information.
}
\description{
Extracts current status data from a ws_monitor object. Returned data include the following:
\itemize{
\item{\code{monitorID} - the PWFSL monitor ID}
\item{\code{lastValidTime} - UTC POSIXct time corresponding to the last valid pm25 datum}
\item{\code{lastValidLocalTimestamp} - ASCII version of \code{lastValidTime} in monitor-local timezone}
\item{\code{processingTime} - UTC POSIXct when the function is run (\emph{i.e.} 'now')}
\item{\code{latency} - (difference between processingTime -1 hr) and lastValidTime, floored to the hour}
\item{\code{PM2.5_latest_nowcast} - NowCast value at lastValidTime}
\item{\code{PM2.5_latest_1} - PM2.5 value at lastValidTime (should never be null)}
\item{\code{PM2.5_latest_3} - mean of the three hours preceeding lastValidTime}
\item{\code{PM2.5_yesterday} - local time midnight-to-midnight 24-hour mean for the day prior to processingTime}
}
The three-hour average will remove missing values and may represent an average of 1-3 hours.

The yesterday average will remove up to six missing values, returning \code{NA} if more than six hours of yesterday's data are missing.

\strong{NOTE:} \code{PM2.5_yesterday} represents yesterday relative to \code{processingTime} -- \emph{i.e.} actually yesterday.
The various 'latest' values describe the most recent values, whenever they occurred.
}
\details{
Data are assigned to the beginning of the hour they represent. So a PM2.5 value assigned to 2pm
will represent data averaged over the period 14:00 - 14:59. This is in keeping with a day representing
00:00 - 23:59 and a month representing the beginning of the month until the end of the month.

Because of this, data for 2pm should never be available until just after 3pm. If it is currently 3:15pm
then we need to subtract 1 hour from the current \code{processingTime} before subtracting the \code{lastValidTime}
to generate the \code{latency}.
}
\examples{
\dontrun{
wa <- loadLatest() \%>\% monitor_subset(stateCodes = 'WA')
wa_current <- monitor_currentData(wa)
}
}
\keyword{ws_monitor}
