% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_dailyStatistic.R
\name{monitor_dailyStatistic}
\alias{monitor_dailyStatistic}
\title{Calculate Daily Statistics for each Monitor in a ws_monitor Object}
\usage{
monitor_dailyStatistic(ws_monitor, FUN = get("mean"), dayStart = "midnight",
  na.rm = TRUE)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{FUN}{function used to collapse a day's worth of data into a single number for each monitor in the ws_monitor object}

\item{dayStart}{one of \code{sunset|midnight|sunrise}}

\item{na.rm}{logical value indicating whether NA values should be ignored (\code{na.rm = TRUE} --> ignore NAs)}
}
\value{
ws_monitor object with daily statistics
}
\description{
Calculates daily statistics for each monitor in the \code{ws_monitor} data list.
}
\details{
Sunrise and sunset times are calculated based on the first monitor encountered.
This should be accurate enough for all use cases involving co-located monitors. Monitors
from different regions should have daily statistics calculated separately.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke')
airnow <- airnow_load(20150801, 20150831)
WA_smoky <- monitor_subset(airnow, stateCodes='WA', vlim=c(55,Inf))
WA_smoky_dailyMean <- monitor_dailyStatistic(WA_smoky, FUN=get('mean'), dayStart='midnight')
monitor_timeseriesPlot(WA_smoky_dailyMean, type='s')
}
}
\keyword{ws_monitor}

