% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_dailyThreshold.R
\name{monitor_dailyThreshold}
\alias{monitor_dailyThreshold}
\title{Calculate Daily Counts of Values At or Above a Threshold}
\usage{
monitor_dailyThreshold(ws_monitor, threshold = "unhealthy",
  dayStart = "midnight", minHours = 0, na.rm = TRUE)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{threshold}{AQI level name (e.g. \code{"unhealthy"}) or numerical threshold at or above which a measurement is counted}

\item{dayStart}{one of \code{"sunset|midnight|sunrise"}}

\item{minHours}{minimum number of hours to be reported}

\item{na.rm}{logical value indicating whether NA values should be ignored}
}
\value{
ws_monitor object with daily count of hours at or above threshold
}
\description{
Calculates the number of hours per day each monitor in a ws_monitor object was at or above a given threshold
}
\details{
\strong{NOTE:} The returned counts include values at OR ABOVE the given threshold; this applies to both categories and values. 
For example, passing a \code{threshold} argument = "unhealthy" will return a daily count of values that are unhealthy, 
very unhealthy, or extreme (i.e. >= 55.5), as will passing a \code{threshold} argument = 55.5.

AQI levels for \code{threshold} argument = one of \code{"good|moderate|USG|unhealthy|very unhealthy|extreme"}

Sunrise and sunset times are calculated based on the first monitor encountered.
This should be accurate enough for all use cases involving co-located monitors. Monitors
from different regions should have daily statistics calculated separately.
}
\examples{
\dontrun{
AirNow <- airnow_load(20150801, 20150930)

WA_IDS <- AirNow$meta$monitorID[AirNow$meta$stateCode == 'WA']
WA <- monitor_subset(AirNow, monitorIDs=WA_IDS)
monitor_leaflet(WA)

# Pull out specific locations
Twisp <- monitor_subset(WA, monitorIDs='530470009')
Winthrop <- monitor_subset(WA, monitorIDs='530470010')
Omak <- monitor_subset(WA, monitorIDs='530470013')

# Generate 
Twisp_daily <- monitor_dailyThreshold(Twisp, 89, dayStart='midnight', minHours=1)
Winthrop_daily <- monitor_dailyThreshold(Winthrop, 89, dayStart='midnight', minHours=1)
Omak_daily <- monitor_dailyThreshold(Omak, 89, dayStart='midnight', minHours=1)

# Plot
cols <- c(adjustcolor('orange',0.6),
          adjustcolor('red',0.6),
          adjustcolor('blue',0.6))
legend <- c('Twisp','Winthrop','Omak')
monitor_timeseriesPlot(Twisp_daily, pch=16, col=cols[1])
monitor_timeseriesPlot(Winthrop_daily, pch=16, col=cols[2], add=TRUE)
monitor_timeseriesPlot(Omak_daily, pch=16, col=cols[3], add=TRUE)
legend('topright', legend=legend, col=cols, pch=16)
title('Hours per day Above Threshold (PM 2.5 = 89)')
}
}
\keyword{ws_monitor}

