% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_distance.R
\name{monitor_distance}
\alias{monitor_distance}
\title{Calculate Distances from Target Location}
\usage{
monitor_distance(ws_monitor, lon, lat)
}
\arguments{
\item{ws_monitor}{object}

\item{lon}{longitude of the location of interest}

\item{lat}{latitude of the location of interest}
}
\value{
vector of of distances (km)
}
\description{
This function returns the distances (km) between monitoring sites and
the location of interest. These distances can be used to create a mask identifying
monitors within a certain radius of the target location.
}
\examples{
\dontrun{
airnow <- airnow_load(2014091300, 2014101023)
distance <- monitor_distance(airnow, -120.604, 38.782)
closestIndex <- which(distance == min(distance))
distance[closestIndex]
airnow$meta[closestIndex,]
}
}
\seealso{
distance
}
\keyword{monitor}

