% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_dygraph.R
\name{monitor_dygraph}
\alias{monitor_dygraph}
\title{Dygraph Interactive Timeseries}
\usage{
monitor_dygraph(ws_monitor, title = "title", ylab = "PM2.5 Concentration",
  dateWindow = NULL, rollPeriod = 1, showLegend = TRUE)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{title}{title text}

\item{ylab}{title for the y axis}

\item{dateWindow}{initial time range (POSIXct)}

\item{rollPeriod}{rolling mean to be applied to the data}

\item{showLegend}{logical to toggle display of the legend}
}
\value{
Initiates the interactive dygraph plot in RStudio's 'Viewer' tab.
}
\description{
This function creates interactive graphs that will be displayed in RStudio's 'Viewer' tab.
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
King_Fire <- monitor_subsetByDistance(airnow, lon=-120.604, lat=38.782, radius=50)
monitor_dygraph(King_Fire, title='KingFire/California/2014', rollPeriod=3)
}
}
\keyword{ws_monitor}

