% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_getDailyMean.R
\name{monitor_getDailyMean}
\alias{monitor_getDailyMean}
\title{Calculate Daily Means for a Monitor Object}
\usage{
monitor_getDailyMean(ws_monitor, monitorID = NULL, startdate = NULL,
  enddate = NULL)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{monitorID}{Monitor ID of interest.}

\item{startdate}{Desired start date (integer or character in Ymd format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}
}
\value{
A vector of daily means for
}
\description{
Calculates and returns daily means for a monitor.
}
\examples{
monitor_getDailyMean(PWFSLSmoke::Carmel_Valley, startdate = "2016-08-08")
}
