% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_hourByDayStackedBarPlot.R
\name{monitor_hourByDayStackedBarPlot}
\alias{monitor_hourByDayStackedBarPlot}
\title{Hour by Day Bar Plot}
\usage{
monitor_hourByDayStackedBarPlot(ws_monitor, monitorID = NULL, stacks = 4)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}

\item{stacks}{number of stacks in a 24 hour period}
}
\description{
A bar plot.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150810)
monitor <- ws_monitor$meta$monitorID[1]
monitor_hourByDayStackedBarPlot(ws_monitor, monitor)
}
}

