% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_isolate.R
\name{monitor_isolate}
\alias{monitor_isolate}
\title{Isolate Individual Monitors}
\usage{
monitor_isolate(ws_monitor, xlim = NULL, ylim = NULL, tlim = NULL,
  monitorIDs = NULL, stateCodes = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{xlim}{optional longitude lim with lo and hi longitude values}

\item{ylim}{optional latitude lim with lo and hi latitude values}

\item{tlim}{optional time lim with lo and hi time values (POSIXct)}

\item{monitorIDs}{optional vector of monitorIDs}

\item{stateCodes}{optional vector of stateCodes}
}
\value{
A List of isolated \code{ws_monitor} objects.
}
\description{
Filters the incoming ws_monitor object according to the parameters
passed in.  If any parameter is not specified, that parameter will not be used in the filtering.

After filtering, each monitorID found in the \code{ws_monitor} object is extracted
and its \code{data} dataframe is restricted to the times from when that monitor first
datapoint until its last datapoint.

This function is useful when \code{ws_monitor} objects are created for
mobile monitors that are deployed to different locations in different years.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke')
airsis <- airsis_load(20140101, 20151231)
monitorList <- monitor_isolate(airsis)
names(monitorList)
}
}
\seealso{
\link{monitor_subset}
}
\keyword{ws_monitor}

