% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_join.R
\name{monitor_join}
\alias{monitor_join}
\title{Stitch Together Monitors with Shared monitorIDs}
\usage{
monitor_join(ws_monitor1 = NULL, ws_monitor2 = NULL, monitorIDs = NULL)
}
\arguments{
\item{ws_monitor1}{\emph{ws_monitor} object}

\item{ws_monitor2}{\emph{ws_monitor} object}

\item{monitorIDs}{vector of shared monitorIDs that are to be joined together}
}
\value{
A \emph{ws_monitor} object with merged timeseries.
}
\description{
For each monitor in \code{monitorIDs}, an attempt is made to merge
the associated data from \code{ws_monitor} and \code{ws_monitor2} and.

This is useful when the same \code{monitorID} appears in different \emph{ws_monitor}
objects representing different time periods.
}
\examples{
\dontrun{
Jul <- monitor_subset(Northwest_Megafires,
                      tlim=c(20150701,20150730),
                      timezone='America/Los_Angeles')
Aug <- monitor_subset(Northwest_Megafires,
                      tlim=c(20150801,20150831),
                      timezone='America/Los_Angeles')
Methow_Valley <- monitor_join(Jul, Aug, monitorIDs=c('530470010','530470009'))
}
}
\keyword{ws_monitor}
