% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_leaflet.R
\name{monitor_leaflet}
\alias{monitor_leaflet}
\title{Leaflet Interactive Map of Monitoring Stations}
\usage{
monitor_leaflet(ws_monitor, slice = get("max"), AQIStyle = "1_3",
  breaks = NULL, radius = 10, opacity = 0.7, providerTiles = NULL,
  popupInfo = c("siteName", "monitorID", "elevation"))
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{slice}{either a time index or a function used to collapse the time axis -- defautls to \code{get('max')}}

\item{AQIStyle}{AQI breaks and colors to use ('1_3'|'8'|'24') -- defaults to '1_3'}

\item{breaks}{7 levels used instead of AQI break levels}

\item{radius}{radius of monitor circles}

\item{opacity}{opacity of monitor circles}

\item{providerTiles}{optional name of leaflet ProviderTiles to use, e.g. "Stamen.Terrain"}

\item{popupInfo}{a vector of column names from ws_monitor$meta to be shown in a popup window}
}
\value{
Initiates the interactive dygraph plot in Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in RStudio's 'Viewer' tab.
Individual monitor timeseries are reduced to
a single value by applying \code{statistic}to the entire timeseries of each monitor with \code{na.rm=TRUE}.
These values are then plotted over a map of the United States.
}
\details{
The colors used for the monitors will be chosen based on the AQIStyle which should match the
width of any rolling mean that was applied to the data. Note that the AQI colors used for 1-hr and 3-hr
rolling means are the same.

If \code{slice} is a function (not a function name) it will be used with argument \code{na.rm=TRUE} to
collapse the time dimension. Thus, user defined functions must accept \code{na.rm} as a parameter.

See \url{https://leaflet-extras.github.io/leaflet-providers/} for a list of "provider tiles".
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
v_low <- AQI$breaks_1_3[4]
CA_unhealthy_monitors <- monitor_subset(airnow, stateCodes='CA', vlim=c(v_low, Inf))
monitor_leaflet(CA_unhealthy_monitors, providerTiles="Stamen.Terrain")
}
}
\keyword{ws_monitor}

