% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_map.R
\name{monitor_map}
\alias{monitor_map}
\title{Create Map of Monitoring Stations}
\usage{
monitor_map(ws_monitor, slice = get("max"), breaks = 7,
  paletteFunc = colorRampPalette(c("gray90", "antiquewhite", "yellow", "red",
  "firebrick")), showCounties = TRUE, showLegend = TRUE, add = FALSE, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{slice}{either a time index or a function used to collapse the time axis -- defautls to \code{get('max')}}

\item{breaks}{set of breaks used to assign colors or a single integer used to provide quantile based breaks}

\item{paletteFunc}{a palette generating function as returned by \code{colorRampPalette}}

\item{showCounties}{logical specifying whether to plot county boundaries -- defaults to \code{FALSE}}

\item{showLegend}{logical specifying whether to add a legend -- defaults to \code{TRUE}}

\item{add}{logical specifying whether to add points to an existing map}

\item{...}{additional arguments passed to points() such as 'col' or 'pch'}
}
\description{
Creates a map of monitoring stations in the US. Individual monitor timeseries are reduced to 
a single value by applying the function passed in as \code{slice} to the entire timeseries of each monitor
with \code{na.rm=TRUE}. These values are then plotted over a map of the United States. All arguments specified in '...' are passed
on to the points() function.

If \code{slice} is an integer, it will be used as an index to pull out a single timestep.

If \code{slice} is a function (not a function name) it will be used with argument \code{na.rm=TRUE} to
collapse the time dimension. Thus, user defined functions must accept \code{na.rm} as a parameter.
}
\details{
Using a single number for the \code{breaks} argument will cause the algorithm to use
quantiles to determine breaks.

You can use AQI colors and 24-hr, daily average breaks by specifying \code{breaks=NULL, paletteFunc=Null}.
}
\examples{
\dontrun{
airnow <- airnow_load(20150901, 20150930)
airnow_conus <- monitor_subset(airnow, stateCodes=CONUS)
monitor_map(airnow_conus)
title('Max PM2.5 Levels in September, 2015')
}
}
\keyword{ws_monitor}

