% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_noDataPlot.R
\name{monitor_noDataPlot}
\alias{monitor_noDataPlot}
\title{Creates Blank Plot with "No Data" Message}
\usage{
monitor_noDataPlot(ws_monitor, monitorID = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the given ws_monitor object}
}
\description{
Creates a blank plot that indicates there is no data for the monitor of interest
}
\keyword{ws_monitor}

