% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_noDataPlot.R
\name{monitor_noDataPlot}
\alias{monitor_noDataPlot}
\title{No Data Plot}
\usage{
monitor_noDataPlot(ws_monitor, monitorID)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}
}
\description{
Plot that indicates there is no data
}

