% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_rollingMeanPlot.R
\name{monitor_rollingMeanPlot}
\alias{monitor_rollingMeanPlot}
\title{Create Individual Value and Rolling Mean Plot}
\usage{
monitor_rollingMeanPlot(ws_monitor, monitorID = NULL, width = 24,
  data.thresh = 75, align = "center")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the ws_monitor object (optional if only one monitor
in the ws_monitor object)}

\item{width}{number of periods to average (e.g. for hourly data, \code{width = 24} plots 24-hour rolling means)}

\item{data.thresh}{minimum number of valid observations required as a percent of \code{width};
NA is returned if insufficicnet valid data to calculate mean}

\item{align}{alignment of averaging window relative to point being calculated; one of \code{"left|center|right"}}
}
\description{
Creates a plot of individual (e.g. hourly) and rolling mean PM2.5 values for a specific monitor.
}
\details{
\code{align = "left"} and \code{align = "right"} calculate backward and forward rolling averages, respectively (e.g. current period and prior/subsequent \code{width - 1} periods)
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150725, 20150805)
monitor <- ws_monitor$meta$monitorID[1]
monitor_rollingMeanPlot(ws_monitor, monitor)
}
}
\keyword{ws_monitor}

