% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_rollingMeanPlot.R
\name{monitor_rollingMeanPlot}
\alias{monitor_rollingMeanPlot}
\title{Hourly and Rolling Mean Plot}
\usage{
monitor_rollingMeanPlot(ws_monitor, monitorID = NULL, width = 24,
  data.thresh = 75, align = "center")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the ws_monitor object (optional
if only one monitor in the ws_monitor object)}

\item{width}{width of window (in hours) used to calculate means}

\item{data.thresh}{minimum number of observations as a percent of width required, or NA is returned}

\item{align}{alignment of averaging window relative to point being calculated ["left"|"center"|"right"]}
}
\description{
A graph showing hourly and rolling mean PM2.5 values for a specific monitor (default = 24-hour).
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150725, 20150805)
monitor <- ws_monitor$meta$monitorID[1]
monitor_rollingMeanPlot(ws_monitor, monitor)
}
}

