% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_scaleData.R
\name{monitor_scaleData}
\alias{monitor_scaleData}
\title{Scale ws_monitor Data}
\usage{
monitor_scaleData(ws_monitor, factor)
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object}

\item{factor}{numeric used to scale the data}
}
\value{
A \emph{ws_monitor} object with scaled data.
}
\description{
Scale the data in a \emph{ws_monitor} object by mutiplying it with \code{factor}.
}
\examples{
library(PWFSLSmoke)

wa <- monitor_subset(Northwest_Megafires, stateCodes='WA')
wa_zero <- monitor_scaleData(wa, 3.4)
}
\keyword{ws_monitor}
