% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_subset.R
\name{monitor_subset}
\alias{monitor_subset}
\title{Create a Subset of a ws_monitor Object}
\usage{
monitor_subset(ws_monitor, xlim = NULL, ylim = NULL, tlim = NULL,
  vlim = NULL, monitorIDs = NULL, stateCodes = NULL,
  dropMonitors = TRUE, timezone = "UTC")
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{xlim}{vector with low and high longitude limit values}

\item{ylim}{vector with low and high latitude limit values}

\item{tlim}{optional vector with start and end times (integer or character representing YYYYMMDD[HH] or \code{POSIXct})}

\item{vlim}{vector with low and high PM2.5 data limit values}

\item{monitorIDs}{vector of monitor IDs used to filter the data}

\item{stateCodes}{vector of state codes used to filter the data}

\item{dropMonitors}{flag specifying whether to remove monitors with no data}

\item{timezone}{Olson timezone passed to \code{link{parseDatetime}} when parsing numeric \code{tlim}}
}
\value{
ws_monitor object as a subset of the input ws_monitor object
}
\description{
Creates a subset of a ws_monitor object based on one or more optional input parameters.
If any input parameter is not specified, that parameter will not be used to subset the ws_monitor object.
}
\details{
By default, this function will return a ws_monitor object whose 'data' dataframe has the
    same number of columns as the incoming dataframe, unless any of the columns consist of all \code{NA}s, 
    in which case such columns will be removed (e.g. if there are no valid data for a specific monitor 
    after subsetting by tlim or vlim).
    If \code{dropMonitors=FALSE}, columns that consist of all \code{NA}s will be retained.
}
\examples{
\dontrun{
airnow <- airnow_load(20140913, 20141010)
xlim <- c(-124.801, -122.801)
ylim <- c(47.004, 48.404)
Olympic_Peninsula <- monitor_subset(airnow, xlim, ylim)
monitor_map(Olympic_Peninsula)
}
}
\keyword{ws_monitor}

