% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_subsetBy.R
\name{monitor_subsetBy}
\alias{monitor_subsetBy}
\title{Subset Monitoring Data}
\usage{
monitor_subsetBy(ws_monitor, filter)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{filter}{a filter to use on the ws_monitor object}
}
\value{
A ws_monitor object with a subset of the input ws_monitor object.
}
\description{
The incoming ws_monitor object is filtered according to \code{filter}.
Either meta data or actual data can be filtered.
}
\examples{
\dontrun{
# We will load airnow monitors for all of 2015 and map all monitors in the Pacific timezone
# that have measured hazardous pm2.5 levels.
airnow <- airnow_load(20150101, 20151231)
mySubset <- monitor_subsetBy(airnow, timezone == 'America/Los_Angeles')
hazardousSubset <- monitor_subsetBy(mySubset, data > AQI$breaks_24[6])
monitorLeaflet(hazardousSubset,
               popupInfo=c('siteName', 'agencyName', 'elevation','monitorID'))
}
}
\keyword{ws_monitor}

