% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_subsetBy.R
\name{monitor_subsetBy}
\alias{monitor_subsetBy}
\title{Subset Monitoring Data}
\usage{
monitor_subsetBy(ws_monitor, filter)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{filter}{a filter to use on the ws_monitor}
}
\value{
monitoring data list filtered as specified
}
\description{
The incoming monitoring data list is filtered according to the filter 
passed in. Either meta data or actual data can be filtered.
}
\examples{
\dontrun{
# We will load wrcc monitors from the past 6 years and map all monitors 
# that have measured hazardous pm2.5 levels. 
wrcc <- wrcc_load(20100101, 20160101)
mySubset <- monitor_subsetBy(wrcc, timezone == 'America/Los_Angeles')
hazardousSubset <- monitor_subsetBy(mySubset, data > AQI$breaks_24[6])
monitorInteractiveMap(hazardousSubset,
                popupInfo=c('siteName', 'agencyName', 'elevation','monitorID'))
}
}
\keyword{ws_monitor}

