% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_subsetByDistance.R
\name{monitor_subsetByDistance}
\alias{monitor_subsetByDistance}
\title{Subset Monitor Data by Distance from Location}
\usage{
monitor_subsetByDistance(ws_monitor, lon = NULL, lat = NULL, radius = 300)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{lon}{target longitude from which the radius will be calculated}

\item{lat}{target latitude from which the radius will be calculated}

\item{radius}{distance (km) of radius from target location -- default=300}
}
\value{
\code{ws_monitor} data list containing monitors near a location
}
\description{
This function subsets a monitoring dataList to include only those
monitors within a certain radius of a target location.
If no monitors fall within the specified radius, \code{ws_monitor$data} and
\code{ws_monitor$meta} are set to \code{NULL}.
}
\examples{
\dontrun{
setSmokeDataDir('~/Data/Smoke')
airnow <- airnow_load(20140913, 20141010)
King_Fire <- monitor_subsetByDistance(airnow, lon=-120.604, lat=38.782, radius=50)
monitor_leaflet(King_Fire)
}
}
\seealso{
monitorDistance
}
\keyword{monitor}

