% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_subsetMeta.R
\name{monitor_subsetMeta}
\alias{monitor_subsetMeta}
\title{Subset ws_monitor Object 'meta' Dataframe}
\usage{
monitor_subsetMeta(meta, xlim = NULL, ylim = NULL, stateCodes = NULL,
  monitorIDs = NULL)
}
\arguments{
\item{meta}{ws_monitor object 'meta' dataframe}

\item{xlim}{optional vector with low and high longitude values}

\item{ylim}{optional vector with low and high latitude values}

\item{stateCodes}{optional vector of stateCodes}

\item{monitorIDs}{optional vector of monitorIDs}
}
\value{
ws_monitor object 'meta' dataframe, or \code{NULL} if filtering removes all monitors
}
\description{
Subsets a ws_monitor object's 'meta' dataframe by removing any monitors that
    lie outisde the geographical ranges specified (i.e. outside of the given longitudes and
    latitudes and/or states) and that are not mentioned in the list of monitorIDs.
}
\details{
Longitudes must be specified in the domain [-180,180].
}
\keyword{ws_monitor}

