% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_subsetMeta.R
\name{monitor_subsetMeta}
\alias{monitor_subsetMeta}
\title{Subset ws_monitoring Object 'meta' Dataframe}
\usage{
monitor_subsetMeta(meta, xlim = NULL, ylim = NULL, stateCodes = NULL,
  monitorIDs = NULL)
}
\arguments{
\item{meta}{'meta' dataframe from a \code{ws_monitor} object}

\item{xlim}{optional longitude lim with lo and hi longitude values}

\item{ylim}{optional latitude lim with lo and hi latitude values}

\item{stateCodes}{optional vector of stateCodes}

\item{monitorIDs}{optional vector of monitorIDs}
}
\value{
monitoring 'meta' dataframe or \code{NULL} if filtering removes all monitors
}
\description{
The incoming 'meta' dataframe is filtered to remove any monitors that
    lie outisde the specified range of longitudes and latitudes and that are not mentioned
    in the list of monitorIDs.
    
If any parameter is not specified, that parameter will not be used in the filtering
}
\details{
Longitudes must be specified in the domain [-180,180].
}
\keyword{monitor}

