% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_timeOfDayMultiPiePlot.R
\name{monitor_timeOfDayMultiPiePlot}
\alias{monitor_timeOfDayMultiPiePlot}
\title{Time of Day Multi Pie Plot}
\usage{
monitor_timeOfDayMultiPiePlot(ws_monitor, monitorID = NULL)
}
\arguments{
\item{ws_monitor}{a ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}
}
\description{
Pie plot that shows 9 PM 2.5 levels based on time of day.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150810)
monitor <- ws_monitor$meta$monitorID[1]
monitor_timeOfDayMultiPiePlot(ws_monitor, monitor)
}
}

