% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_timeOfDaySpaghettiPlot.R
\name{monitor_timeOfDaySpaghettiPlot}
\alias{monitor_timeOfDaySpaghettiPlot}
\title{Time of Day Spaghetti Plot}
\usage{
monitor_timeOfDaySpaghettiPlot(ws_monitor, monitorID)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{id for a specific monitor in the ws_monitor}
}
\description{
Spaghetti Plot that shows PM 2.5 levels by hours for 10 days.
The plot also has takes and draws a mean.
}
\examples{
\dontrun{
ws_monitor <- wrcc_load(20150801, 20150810)
monitor <- ws_monitor$meta$monitorID[1]
monitor_timeOfDaySpaghettiPlot(ws_monitor, monitor)
}
}

