% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_timeseriesPlot.R
\name{monitor_timeseriesPlot}
\alias{monitor_timeseriesPlot}
\title{Timeseries Plot for Monitoring Data}
\usage{
monitor_timeseriesPlot(ws_monitor, AQIStyle = FALSE, useGMT = FALSE,
  shadedNight = FALSE, add = FALSE, ...)
}
\arguments{
\item{ws_monitor}{data list of class \code{ws_monitor}}

\item{AQIStyle}{boolean flag specifying use of AQI 24-hr, daily average breaks and colors}

\item{useGMT}{If TRUE, \code{monitor_timeseriesPlot} will display time as
GMT. Otherwise, it will attempt to use local time, provided all monitors
are in the same timezone. There will be an error if \code{useGMT} is FALSE but
monitors are in multiple timezones.}

\item{shadedNight}{Shade nighttime values if TRUE, provided \code{useGMT} is FALSE.
This is because we only allow shading for multiple monitors if they are in
the same timezone; in this case, they have approximately the same sunrise and
sunset times.}

\item{add}{A logical whether you want to add the plot on top of the other timeseries plot}

\item{...}{Any additional graphical parameters that can be passed in to the function}
}
\description{
Plots a time series optionally using AQI colors and breaks.
}
\examples{
\dontrun{
airnow <- airnow_load(20150801, 20150831)
Roseburg <- monitor_subset(airnow, monitorIDs=c('410190002'))
monitor_timeseriesPlot(Roseburg)
}
}
\keyword{ws_monitor}

