% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monitor_timeseriesPlot.R
\name{monitor_timeseriesPlot}
\alias{monitor_timeseriesPlot}
\title{Create a Timeseries Plot of ws_monitor Object PM2.5 Data}
\usage{
monitor_timeseriesPlot(ws_monitor, AQIStyle = "", useGMT = FALSE,
  shadedNight = FALSE, add = FALSE, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{AQIStyle}{AQI styling, one of \code{"1_3|8|24"}}

\item{useGMT}{If TRUE, \code{monitor_timeseriesPlot} will display time as
GMT. Otherwise, it will attempt to use local time, provided all monitors
are in the same timezone. There will be an error if \code{useGMT} is FALSE but
monitors are in multiple timezones.}

\item{shadedNight}{Shade nighttime values if TRUE, provided \code{useGMT} is FALSE.
This is because we only allow shading for multiple monitors if they are in
the same timezone; in this case, they have approximately the same sunrise and
sunset times.}

\item{add}{A logical specifying whether you want to add the plot on top of an existing time series plot}

\item{...}{Any additional graphical parameters that can be passed in to the function}
}
\description{
Creates a time series plot of PM2.5 data from a ws_monitor object (see note below). Optional arguments
color code by AQI index, add shading to indicate nighttime, and adjust the time display (GMT vs. local).
}
\note{
Remember that a ws_monitor object can contain data from more than one monitor, and thus, this function may produce
a time series of data from multiple monitors. To plot a time series of an individual monitor's data, use the monitor_subset
function (see example below).
}
\examples{
\dontrun{
airnow <- airnow_load(20150801, 20150831)
Roseburg <- monitor_subset(airnow, monitorIDs=c('410190002'))
monitor_timeseriesPlot(Roseburg)
}
}
\keyword{ws_monitor}

