% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_trim.R
\name{monitor_trim}
\alias{monitor_trim}
\title{Trim ws_monitor Time Axis to Remove NA Periods From Beginning and End}
\usage{
monitor_trim(ws_monitor)
}
\arguments{
\item{ws_monitor}{ws_monitor object}
}
\value{
ws_monitor object
}
\description{
Trims the time axis of a ws_monitor object to exclude timestamps prior to the first and
after the last valid datapoint for any monitor.
}
\keyword{ws_monitor}

