% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_dailyBarPlot.R
\name{monitorPlot_dailyBarPlot}
\alias{monitorPlot_dailyBarPlot}
\title{Create Daily Bar Plot}
\usage{
monitorPlot_dailyBarPlot(ws_monitor, monitorID = NULL, tlim = NULL,
  minHours = 20, title = NULL, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the given ws_monitor object (optional
if only one monitor in the ws_monitor object)}

\item{tlim}{time limit for barplot}

\item{minHours}{minimum number of valid data hours required to calculate each daily average}

\item{title}{plot title}

\item{...}{additional arguments to be passed to barplot()}
}
\description{
Creates a bar plot showing daily average PM 2.5 values for a specific monitor in a ws_monitor object.
Each bar is colored according to its AQI category.
}
\examples{
\dontrun{
airnow <- airnow_load(20150701, 20150831)
title <- "Daily Average PM2.5 for Colleville, WA"
monitorPlot_dailyBarPlot(airnow, monitorID="530650004", title=title)
}
}
\keyword{ws_monitor}

