% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_dailyBarplot.R
\name{monitorPlot_dailyBarplot}
\alias{monitorPlot_dailyBarplot}
\title{Create Daily Bar Plot}
\usage{
monitorPlot_dailyBarplot(ws_monitor, monitorID = NULL, tlim = NULL,
  minHours = 20, gridPos = "", gridCol = "black", gridLwd = 0.5,
  gridLty = "solid", ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the given ws_monitor object (optional
if only one monitor in the ws_monitor object)}

\item{tlim}{optional vector with start and end times (integer or character representing YYYYMMDD[HH])}

\item{minHours}{minimum number of valid data hours required to calculate each daily average}

\item{gridPos}{position of grid lines either 'over', 'under' ('' for no grid lines)}

\item{gridCol}{color of grid lines (see graphical parameter 'col')}

\item{gridLwd}{line width of grid lines (see graphical parameter 'lwd')}

\item{gridLty}{type of grid lines (see graphical parameter 'lty')}

\item{...}{additional arguments to be passed to barplot()}
}
\description{
Creates a bar plot showing daily average PM 2.5 values for a specific monitor
in a ws_monitor object. Each bar is colored according to its AQI category.

This function is a wrapper around \code{base::barplot} and any arguments to that 
function may be used.

Each 'day' is the midnight-to-midnight period in the monitor local timezone.
When \code{tlim} is used, it is converted to the monitor local timezone.
}
\examples{
\dontrun{
airnow <- airnow_load(20150701, 20150831)
main <- "Daily Average PM2.5 for Colleville, WA"
monitorPlot_dailyBarplot(airnow, monitorID="530650004", main=main)
}
}
\keyword{ws_monitor}

