% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_dailyBarplotFacets.R
\name{monitorPlot_dailyBarplotFacets}
\alias{monitorPlot_dailyBarplotFacets}
\title{Create Daily AQI Barplots of Multiple Monitors}
\usage{
monitorPlot_dailyBarplotFacets(ws_monitor, monitorIDs = NULL,
  title = "Daily AQI", smooth_func = monitor_nowcast,
  smooth_args = list(), smooth_style = c("bars", "points"),
  smooth_name = "Hourly NowCast", theme = theme_monitors(),
  date_format = "\%b \%d", ...)
}
\arguments{
\item{ws_monitor}{\code{ws_monitor} object}

\item{monitorIDs}{vector of monitor IDs in \code{ws_monitor} to subset and plot (if \code{NULL}, all monitors are plotted)}

\item{title}{plot title}

\item{smooth_func}{\code{PWFSLSmoke} smoothing function (set to \code{NULL} to exclude smoothed values)}

\item{smooth_args}{list of additional arguments to the smoothing function}

\item{smooth_style}{how smoothed values are plotted}

\item{smooth_name}{name of smoothing method used in legend}

\item{theme}{ggplot2 \code{\link[ggplot2]{theme}} controlling plotting options (see \code{\link[PWFSLSmoke]{theme_monitors}} for default)}

\item{date_format}{formatting code for x-axis dates (see \code{\link[base]{strftime}} for codes)}

\item{...}{additional arguments, such as \code{tlim}, passed to \code{\link[PWFSLSmoke]{monitor_subset}}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Plots a facetted set of daily-averaged air quality data for multiple monitoring stations.
}
\examples{
\dontrun{
portland_monitors <- Northwest_Megafires$meta$monitorID[grepl('Portland', Northwest_Megafires$meta$siteName)]
monitorPlot_dailyBarplotFacets(Northwest_Megafires, monitorIDs=portland_monitors, tlim=c(20150815,20150901), timezone='US/Pacific')
}
}
\keyword{ws_monitor}
