% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_hourlyBarplot.R
\name{monitorPlot_hourlyBarplot}
\alias{monitorPlot_hourlyBarplot}
\title{Create Hourly Barplot}
\usage{
monitorPlot_hourlyBarplot(ws_monitor, monitorID = NULL, tlim = NULL,
  localTime = TRUE, style = "AirFire", shadedNight = TRUE, gridPos = "",
  gridCol = "black", gridLwd = 0.5, gridLty = "solid", dayLwd = 2,
  hourLwd = 1, hourInterval = 6, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for a specific monitor in the given ws_monitor object (optional
if only one monitor in the ws_monitor object)}

\item{tlim}{time limit for barplot}

\item{localTime}{logical specifying whether \code{tlim} is in UTC or local time}

\item{style}{named style specification ('AirFire')}

\item{shadedNight}{add nighttime shading}

\item{gridPos}{position of grid lines either 'over', 'under' ('' for no grid lines)}

\item{gridCol}{grid color}

\item{gridLwd}{grid line width}

\item{gridLty}{grid line type}

\item{dayLwd}{day boundary line width (set to 0 to omit day lines)}

\item{hourLwd}{hour boundary line width (set to 0 to omit  hour lines)}

\item{hourInterval}{interval for hour boundary lines}

\item{...}{additional arguments to be passed to barplot()}
}
\description{
Creates a bar plot showing hourly PM 2.5 values for a specific monitor in a ws_monitor object.
Colors are assigned to one of the following styles:
\itemize{
\item{\code{AirFire} -- hourly values colored by AQI 24-hour breaks}
}
}
\examples{
\dontrun{
airnow <- airnow_load(20150701, 20150710)
title <- "Hourly PM2.5 for Colleville, WA"
monitorPlot_hourlyBarplot(airnow, monitorID="530650004")
}
}
\keyword{ws_monitor}

