% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorPlot_timeseries.R
\name{monitorPlot_timeseries}
\alias{monitorPlot_timeseries}
\title{Create a Timeseries Plot of ws_monitor Object PM2.5 Data}
\usage{
monitorPlot_timeseries(ws_monitor, monitorID = NULL, tlim = NULL,
  localTime = TRUE, style = "Airfire", shadedNight = FALSE, add = FALSE,
  gridPos = "", gridCol = "black", gridLwd = 1, gridLty = "solid",
  dayLwd = 0, hourLwd = 0, hourInterval = 6, ...)
}
\arguments{
\item{ws_monitor}{ws_monitor object}

\item{monitorID}{monitor ID for one or more monitor in the ws_monitor object}

\item{tlim}{time limits}

\item{localTime}{create plot in local time; applies to tlim as well}

\item{style}{AQI styling, one of \code{'Airfire| | '}... TODO: Fill in other options}

\item{shadedNight}{Shade nighttime values if TRUE, provided all monitors in the same time zone.}

\item{add}{A logical specifying whether you want to add the data points on top of an existing time series plot}

\item{gridPos}{position of grid lines either 'over', 'under' ('' for no grid lines)}

\item{gridCol}{grid line color}

\item{gridLwd}{grid line width}

\item{gridLty}{grid line type}

\item{dayLwd}{day marker line width}

\item{hourLwd}{hour marker line width}

\item{hourInterval}{interval for grid (max=12)}

\item{...}{additional arguments to be passed to points()}
}
\description{
Creates a time series plot of PM2.5 data from a ws_monitor object (see note below). Optional arguments
color code by AQI index, add shading to indicate nighttime, and adjust the time display (GMT vs. local).
}
\note{
Remember that a ws_monitor object can contain data from more than one monitor, and thus, this function may produce
a time series of data from multiple monitors. To plot a time series of an individual monitor's data, use the monitor_subset
function (see example below).
}
\examples{
\dontrun{ 
airnow <- airnow_load(20150801, 20150831)
Roseburg <- monitor_subset(airnow, monitorIDs=c('410190002'))
monitorPlot_timeseries(Roseburg)
}
}
\keyword{ws_monitor}

