% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke.R
\docType{data}
\name{Northwest_Megafires}
\alias{Northwest_Megafires}
\title{Northwest Megafires Example Dataset}
\format{A list with two elements}
\description{
In the summer of 2015 Washington state had several catastrophic wildfires that led
to many days of heavy smoke in eastern Washington, Oregon and northern Idaho.
The Northwest_Megafires dataset contains AirNow ambient monitoring data for the
Pacific Northwest from May 31 through November 01, 2015 (UTC). Data are stored
as a \emph{ws_monitor} object and are used in many examples in the package
documentation.
}
\details{
Northwest_Megafires example dataset
}
\keyword{datasets}
