% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseDatetime.R
\name{parseDatetime}
\alias{parseDatetime}
\title{Parse Datetime Strings}
\usage{
parseDatetime(datetime, timezone = "UTC")
}
\arguments{
\item{datetime}{character or integer datetimes in YYYYMMDD[HHMMSS] format}

\item{timezone}{Olson timezone at the location of interest}
}
\value{
POSIXct datetimes.
}
\description{
8-, 10-, 12- and 14-digit formats are understood, e.g: 20150721 to 20150721000000.
Integers will be converted to character before parsing.
}
\examples{
starttime <- parseDatetime(2015080718)
}

