% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWFSLSmoke-deprecated.R
\name{PWFSLSmoke-deprecated}
\alias{PWFSLSmoke-deprecated}
\alias{PWFSLSmoke-defunct}
\alias{monitorDygraph}
\alias{monitorEsriMap}
\alias{monitorLeaflet}
\alias{monitorMap}
\alias{monitorMap_performance}
\alias{monitorGoogleMap}
\title{Deprecated functions in PWFSLSmoke}
\usage{
monitorGoogleMap(ws_monitor, slice = get("max"),
  breaks = AQI$breaks_24, colors = AQI$colors, width = 640,
  height = 640, centerLon = NULL, centerLat = NULL, zoom = NULL,
  maptype = "roadmap", grayscale = FALSE, map = NULL, ...)

monitorDygraph(ws_monitor, title = "title",
  ylab = "PM2.5 Concentration", tlim = NULL, rollPeriod = 1,
  showLegend = TRUE)

monitorEsriMap(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, width = 640, height = 640, centerLon = NULL,
  centerLat = NULL, zoom = NULL, maptype = "worldStreetMap",
  grayscale = FALSE, mapRaster = NULL, cex = par("cex") * 2,
  pch = 16, ...)

monitorLeaflet(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, labels = AQI$names,
  legendTitle = "Max AQI Level", radius = 10, opacity = 0.7,
  maptype = "terrain", popupInfo = c("siteName", "monitorID",
  "elevation"))

monitorMap(ws_monitor, slice = get("max"), breaks = AQI$breaks_24,
  colors = AQI$colors, pch = par("pch"), cex = par("cex"),
  stateCol = "grey60", stateLwd = 2, countyCol = "grey70",
  countyLwd = 1, add = FALSE, ...)

monitorMap_performance(predicted, observed, threshold = AQI$breaks_24[3],
  cex = par("cex"), sizeBy = NULL, colorBy = "heidikeSkill",
  breaks = c(-Inf, 0.5, 0.6, 0.7, 0.8, Inf),
  paletteFunc = grDevices::colorRampPalette(RColorBrewer::brewer.pal(length(breaks),
  "Purples")[-1]), showLegend = TRUE, legendPos = "topright",
  stateCol = "grey60", stateLwd = 2, countyCol = "grey70",
  countyLwd = 1, add = FALSE, ...)
}
\description{
These functions are in the process of being removed from the PWFSLSmoke
package, and may not appear in future versions. When available, alternative
functions will be listed.
}
\details{
Deprecations fall into the following categories:

\itemize{
  \item Silent: These functions immediately call another function without any
  warning to the user.
  \item Soft: These functions throw a warning when called, then call the
  correct function, if available.
  \item Hard: These functions will throw an error when called, stopping
  execution. If an alternative function exists, it will be listed.
}
}
\section{Silent Deprecations}{


\tabular{rl}{
 \code{\link{monitorDygraph}} \tab Use \code{\link{monitor_dygraph}}\cr
 \code{\link{monitorEsriMap}} \tab Use \code{\link{monitor_esriMap}}\cr
 \code{\link{monitorLeaflet}} \tab Use \code{\link{monitor_leaflet}}\cr
 \code{\link{monitorMap}} \tab Use \code{\link{monitor_map}}\cr
 \code{\link{monitorMap_performance}} \tab Use \code{\link{monitor_performanceMap}}\cr
}
}

\section{Hard Deprecations}{


\tabular{rl}{
 \code{\link{monitorGoogleMap}} \tab Use \code{\link{monitor_esriMap}}\cr
}
}

\keyword{internal}
