% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_getHighlightDates.R
\name{raw_getHighlightDates}
\alias{raw_getHighlightDates}
\title{Return Day Stamps for Values Above a Threshold}
\usage{
raw_getHighlightDates(df, dataVar, tzone = NULL, highlightRange = c(1e+12,
  Inf))
}
\arguments{
\item{df}{dataframe with \code{datetime} column in GMT}

\item{dataVar}{variable to be evaluated}

\item{tzone}{timezone where data were collected}

\item{highlightRange}{range of values of to be highlighted}
}
\description{
Returna list of dates in YYYYMMDD format where
the dataVar is within \code{highlightRange}.
}
\examples{
\dontrun{
highlightRange <- c(50,Inf)
dataVar <- 'ConcHr'
}
}

