% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_monitors.R
\name{theme_monitors}
\alias{theme_monitors}
\title{Theme for Plotting Multiple Monitors}
\usage{
theme_monitors(title_size = 14, axis_title_size = 12,
  axis_label_size = 10, facet_title_size = 12, legend_title_size = 12,
  legend_label_size = 10, legend_position = "bottom")
}
\arguments{
\item{title_size}{size of plot title}

\item{axis_title_size}{size of axis titles}

\item{axis_label_size}{size of axis labels}

\item{facet_title_size}{size of facet titles}

\item{legend_title_size}{size of legend titles}

\item{legend_label_size}{size of legend item labels}

\item{legend_position}{legend position}
}
\value{
\code{ggplot2::theme}
}
\description{
A theme used by \code{monitorPlot_dailyBarplotFacets}.
}
