% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createDataDataframe.R
\name{wrcc_createDataDataframe}
\alias{wrcc_createDataDataframe}
\title{Create Data Dataframe}
\usage{
wrcc_createDataDataframe(df, meta, verbose = FALSE)
}
\arguments{
\item{df}{single site WRCC dataframe created by wrcc_clustering()}

\item{meta}{WRCC meta dataframe created by wrcc_createMetaDataframe()}

\item{verbose}{logical flag to generate verbose output}
}
\value{
ws_monitor object with a unique `monitorID` for each unique location
}
\description{
Explain this function
}
\keyword{WRCC}

