% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createDataDataframe.R
\name{wrcc_createDataDataframe}
\alias{wrcc_createDataDataframe}
\title{Create WRCC Data Dataframe}
\usage{
wrcc_createDataDataframe(df, meta)
}
\arguments{
\item{df}{single site WRCC dataframe created by wrcc_clustering()}

\item{meta}{WRCC meta dataframe created by wrcc_createMetaDataframe()}
}
\value{
A \code{'data'} dataframe for use in a \code{ws_monitor} object.
}
\description{
After quality control has been applied to an WRCC dataframe,
we can extract the PM2.5 values and store them in a \code{'data'} dataframe
organized as hour-by-deployment (aka hour-by-site).

The first column of the returned dataframe is named \code{'datetime'} and
contains a \code{POSIXct} time in UTC. Additional columns contain data
for each separate deployment of a monitor.
}
\keyword{WRCC}

