% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMetaDataframe.R
\name{wrcc_createMetaDataframe}
\alias{wrcc_createMetaDataframe}
\title{Create WRCC Site Location Metadata Dataframe}
\usage{
wrcc_createMetaDataframe(tbl, unitID = as.character(NA),
  pwfslDataIngestSource = "WRCC", existingMeta = NULL,
  addGoogleMeta = FALSE, addEsriMeta = FALSE)
}
\arguments{
\item{tbl}{single site WRCC tibble after metadata enhancement}

\item{unitID}{character or numeric WRCC unit identifier}

\item{pwfslDataIngestSource}{identifier for the source of monitoring data, e.g. \code{'WRCC', 'WRCC_DUMPFILE'}}

\item{existingMeta}{existing 'meta' dataframe from which to obtain metadata for known monitor deployments}

\item{addGoogleMeta}{logicial specifying wheter to use Google elevation and reverse geocoding services}

\item{addEsriMeta}{logicial specifying wheter to use ESRI elevation and reverse geocoding services}
}
\value{
A \code{meta} dataframe for use in a \emph{ws_monitor} object.
}
\description{
After a WRCC tibble has been enhanced with 
additional columns generated by \code{addClustering} we are ready to 
pull out site information associated with unique deployments.

These will be rearranged into a dataframe organized as deployment-by-property
with one row for each monitor deployment.

This site information found in \code{tbl} is augmented so that we end up with a uniform
set of properties associated with each monitor deployment. The list of
columns in the returned \code{meta} dataframe is:

\preformatted{
> names(p$meta)
 [1] "monitorID"             "longitude"             "latitude"             
 [4] "elevation"             "timezone"              "countryCode"          
 [7] "stateCode"             "siteName"              "agencyName"           
[10] "countyName"            "msaName"               "monitorType"          
[13] "monitorInstrument"     "aqsID"                 "pwfslID"              
[16] "pwfslDataIngestSource" "telemetryAggregator"   "telemetryUnitID"      
}
}
\seealso{
\code{\link{addGoogleMetadata}}

\code{\link{addMazamaMetadata}}
}
\keyword{WRCC}
