% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{wrcc_downloadData}
\alias{wrcc_downloadData}
\title{Download Data from WRCC}
\usage{
wrcc_downloadData(stationID = NULL, startdate = 20100101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  baseUrl = "http://www.wrcc.dri.edu/cgi-bin/wea_list2.pl")
}
\arguments{
\item{stationID}{station identifier (will be upcased)}

\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{baseUrl}{base URL for data queries}
}
\value{
String containing WRCC output.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the WRCC output. 

Monitor stationIDs can be found at http://www.wrcc.dri.edu/cgi-bin/smoke.pl.
}
\examples{
\dontrun{
fileString <- wrcc_downloadData(20150701, 20150930, stationID='SM16')
df <- wrcc_parseData(fileString)
}
}
\references{
\href{http://www.wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
\keyword{WRCC}

