% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{wrcc_downloadData}
\alias{wrcc_downloadData}
\title{Download Data from WRCC}
\usage{
wrcc_downloadData(stationID = NULL, startdate = 20100101,
  enddate = strftime(lubridate::now(), "\%Y\%m\%d", tz = "GMT"),
  baseUrl = "http://www.wrcc.dri.edu/cgi-bin/wea_list2.pl")
}
\arguments{
\item{stationID}{station identifier (will be upcased)}

\item{startdate}{integer or character representing start date as YYYYMMDD (GMT)}

\item{enddate}{integer or character representing end date as YYYYMMDD (GMT)}

\item{baseUrl}{base URL for data queries}
}
\value{
String containing WRCC output.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the WRCC output.

Monitor stationIDs can be found at http://www.wrcc.dri.edu/cgi-bin/smoke.pl.
}
\examples{
\dontrun{
fileString <- wrcc_downloadData('SM16',startdate=20150701,enddate=20150930)
df <- wrcc_parseData(fileString)
}
}
\references{
\href{http://www.wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
\keyword{WRCC}

