% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadLatest.R
\name{wrcc_loadLatest}
\alias{wrcc_loadLatest}
\title{Load Recent WRCC Monitoring Data}
\usage{

  wrcc_loadLatest(baseUrl = "https://haze.airfire.org/monitoring/RData/wrcc_pm25_latest.RData")
}
\arguments{
\item{baseUrl}{The location of the meta and data files}
}
\value{
A ws_monitor object with WRCC data.
}
\description{
The most recent 45 days of WRCC data are updated in real time
at PWFSL and can be loaded with this function.
}
\examples{
\dontrun{
wrcc <- wrcc_loadLatest()
}
}
\keyword{WRCC}

