% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadLatest.R
\name{wrcc_loadLatest}
\alias{wrcc_loadLatest}
\title{Load most recent WRCC monitoring data}
\usage{
wrcc_loadLatest(
  parameter = "PM2.5",
  baseUrl = "https://haze.airfire.org/monitoring/latest/RData",
  dataDir = NULL
)
}
\arguments{
\item{parameter}{Parameter of interest.}

\item{baseUrl}{Base URL for 'daily' AirNow data files.}

\item{dataDir}{Local directory containing 'daily' data files.}
}
\value{
A \emph{ws_monitor} object with WRCC data.
}
\description{
Loads pre-generated .RData files containing the most recent
WRCC data.

If \code{dataDir} is defined, data will be loaded from this local
dirctory. Otherwise, data will be loaded from the monitoring data repository
maintained by PWFSL.

The  files loaded by this function are updated multiple times an hour and
contain data for the previous 10 days.

For daily updates covering the most recent 45 days, use \code{wrcc_loadDaily()}.

For data extended more than 45 days into the past, use \code{wrcc_loadAnnual()}.

WRCC parameters include the following:
\enumerate{
\item{PM2.5}
}

Avaialble RData and associated log files can be seen at:
\href{https://haze.airfire.org/monitoring/WRCC/RData/latest}{https://haze.airfire.org/monitoring/WRCC/RData/latest}
}
\examples{
\dontrun{
wrcc_loadLatest() \%>\%
  monitor_subset(stateCodes=CONUS) \%>\%
  monitor_map()
}
}
\seealso{
\code{\link{wrcc_loadAnnual}}

\code{\link{wrcc_loadDaily}}
}
\keyword{WRCC}
