% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrccDump_createMonitorObject.R
\name{wrccDump_createMonitorObject}
\alias{wrccDump_createMonitorObject}
\title{Ingest WRCC Dump File and Create ws_monitor Object}
\usage{
wrccDump_createMonitorObject(filepath)
}
\arguments{
\item{filepath}{absolute path of the WRCC dump file}
}
\value{
ws_monitor object with a unique `monitorID` for each unique deployment.
}
\description{
Ingests an  WRCC dump file and converts
it into a quality controlled, metadata enhanced \code{ws_monitor} object
ready for use with all \code{monitor_~} functions.

Steps involved include:

\enumerate{
 \item{ingest CSV text}
 \item{parse CSV text}
 \item{apply quality control}
 \item{apply clustering to determine unique deployments}
 \item{enhance metadata to include: elevation, timezone, state, country, site name}
 \item{reshape data into deployment-by-property 'meta' and and time-by-deployment 'data' dataframes}
}
}
\note{
Each dump file must contain data for only one type of monitor, e.g. EBAM or E-Sampler.
}
\seealso{
\code{\link{wrccDump_parseData}}

\code{\link{wrcc_qualityControl}}

\code{\link{addClustering}}

\code{\link{wrcc_createMetaDataframe}}

\code{\link{wrcc_createDataDataframe}}
}
\keyword{WRCC}

